fix(gui): load selected recipe onto the canvas (blank-canvas regression)
Selecting a .yaml recipe in the Files pane only opened the YAML text in the
viewer; the canvas stayed blank. The WorkspaceFilesTree refactor (3a91d7f)
dropped the wiring that loads the rendered figure. Re-add a file-select listener
that calls switchFile() for recipe files (.yaml/.yml), which fetches the render
via api/switch and places it on the canvas (switchFile shows an error toast on
failure -- fail-loud).

- main.tsx: add the file-select -> switchFile listener + store import.
- Extract figrecipeChatAdapter + FIGRECIPE_SYSTEM to bootstrap/chatAdapter.ts
  to keep main.tsx under the 512-line limit.
- Rebuild the static bundle (vite) against current scitex-ui. The served bundle
  was stale (2026-03-22); this brings it current and includes the fix. Build
  uses vite directly: the `tsc` gate currently fails on pre-existing scitex-ui
  type drift (57 errors) unrelated to this change.

Requires scitex-ui#40 (dangling effects.css @import) for the frontend to build.
