+++ src/figrecipe/_django/static/figrecipe/assets/index.js
@@ -273,7 +273,7 @@ ${e.stack||""}`;try{return JSON.stringify(e,null,2)}catch{return String(e)}}getC
 `,r+=`   ${i}
 `}),this.networkErrors.length>0&&this.networkErrors.forEach(i=>{r+=`${i}
 `}),this.consoleLogs.forEach(i=>{const o=i.source?`${i.source} `:"";r+=`${o}${i.args.join(" ")}
-`}),r}getFailedResources(){const e=[];return window.performance&&window.performance.getEntriesByType&&window.performance.getEntriesByType("resource").forEach(s=>{s.responseStatus&&s.responseStatus>=400&&e.push(s.name)}),e}async captureDebugSnapshot(){this.notificationManager.showCameraFlash();const e=await this.captureScreenshotBlob(),n=this.getConsoleLogs();if(!e&&!n){this.notificationManager.showNotification("Copy failed","error"),this.notificationManager.triggerCopyCallback();return}if(e)try{await navigator.clipboard.write([new ClipboardItem({"image/png":e})]),this.notificationManager.showNotification("Screenshot copied - paste now!","success",2500)}catch(s){this.originalConsole.error("[ConsoleCollector] Screenshot clipboard failed:",s)}if(n&&n!=="No console logs captured."){const s=e?3e3:0;await new Promise(r=>setTimeout(r,s));try{await navigator.clipboard.writeText(n),this.notificationManager.showNotification("Console logs copied - paste now!","success")}catch(r){this.originalConsole.error("[ConsoleCollector] Logs clipboard failed:",r)}}else e||this.notificationManager.showNotification("No logs to copy","error");this.notificationManager.triggerCopyCallback()}async captureScreenshotBlob(){try{const e=await navigator.mediaDevices.getDisplayMedia({video:{displaySurface:"browser"},preferCurrentTab:!0,selfBrowserSurface:"include",systemAudio:"exclude"}),n=document.createElement("video");n.srcObject=e,n.muted=!0,await new Promise((o,a)=>{n.onloadedmetadata=()=>{n.play().then(()=>o()).catch(a)},n.onerror=a,setTimeout(()=>a(new Error("Video load timeout")),3e3)}),await new Promise(o=>setTimeout(o,100));const s=document.createElement("canvas");s.width=n.videoWidth,s.height=n.videoHeight;const r=s.getContext("2d");return r?(r.drawImage(n,0,0),e.getTracks().forEach(o=>o.stop()),await new Promise(o=>{s.toBlob(a=>o(a),"image/png")})):(e.getTracks().forEach(o=>o.stop()),null)}catch(e){return e.name!=="NotAllowedError"&&this.originalConsole.error("[ConsoleCollector] Screenshot failed:",e),null}}clearLogs(){this.consoleLogs=[],this.originalConsole.log("[ConsoleCollector] Logs cleared")}}class Bc{constructor(){this.isActive=!1,this.notificationManager=new Yy,this.debugCollector=new Vy,this.overlayManager=new Uy,this.elementScanner=new sa(this.debugCollector,this.notificationManager),this.selectionManager=new Xy(this.elementScanner.getElementBoxMap(),this.debugCollector,this.notificationManager),this.selectionManager.setElementScanner(this.elementScanner),new Gy(this.notificationManager),this.consoleCollector=new qy(this.notificationManager),this.notificationManager.setOnCopyCallback(()=>{this.deactivate()}),this.init()}init(){document.addEventListener("keydown",e=>{const n=e.key.toLowerCase();if(!["Tab","Enter","ArrowUp","ArrowDown","ArrowLeft","ArrowRight"].includes(e.key)){if(e.ctrlKey&&e.shiftKey&&!e.altKey&&n==="i"){e.preventDefault(),e.stopPropagation(),console.log("[ElementInspector] Ctrl+Shift+I pressed - capturing debug snapshot"),this.consoleCollector.captureDebugSnapshot();return}if(e.ctrlKey&&e.altKey&&!e.shiftKey&&n==="i"){e.preventDefault(),this.startSelectionMode();return}if(e.ctrlKey&&!e.altKey&&!e.shiftKey&&n==="i"&&this.isActive){e.preventDefault(),console.log("[ElementInspector] Ctrl+I pressed - loading next batch"),this.elementScanner.loadNextBatch();return}if(e.altKey&&!e.shiftKey&&!e.ctrlKey&&n==="i"){e.preventDefault(),this.toggle();return}if(e.key==="Escape"){this.selectionManager.isActive()?(e.preventDefault(),this.selectionManager.cancelSelectionMode(),this.deactivate()):this.isActive&&(e.preventDefault(),this.deactivate());return}}}),console.log("[ElementInspector] Initialized"),console.log("  Alt+I: Toggle inspector overlay"),console.log("  Ctrl+I: Load next 512 elements (when active)"),console.log("  Ctrl+Alt+I: Rectangle selection mode"),console.log("  Ctrl+Shift+I: Debug snapshot (screenshot + console logs)"),console.log("  Scroll wheel: Cycle through overlapped elements (affects rect selection depth)"),console.log("  Right-click: Copy element debug info"),console.log("  Left-click: Pass through to underlying element"),console.log("  Escape: Deactivate inspector / Cancel selection")}toggle(){this.isActive?this.deactivate():this.activate()}activate(){console.log("[ElementInspector] Activating..."),this.isActive=!0;const e=this.overlayManager.createOverlay();this.elementScanner.scanElements(e),console.log("[ElementInspector] Active - Press Alt+I to deactivate")}deactivate(){console.log("[ElementInspector] Deactivating..."),this.isActive=!1,this.elementScanner.clearElementBoxMap(),this.overlayManager.removeOverlay()}refresh(){this.isActive&&(this.deactivate(),this.activate())}startSelectionMode(){this.isActive||this.activate(),this.selectionManager.startSelectionMode()}}const Uc=window.elementInspector,Qy=Uc instanceof Bc?Uc:new Bc;window.elementInspector=Qy;let Wc;window.addEventListener("resize",()=>{clearTimeout(Wc),Wc=window.setTimeout(()=>{var t;(t=window.elementInspector)!=null&&t.isActive&&window.elementInspector.refresh()},500)});const ra=[];let Me=null;function Zy(){return Me}function qh(t){ra.push(t);const e=localStorage.getItem(t.storageKey);if(e){const n=parseFloat(e);console.debug(`[zoom] restoring ${t.storageKey}: saved=${e}, parsed=${n}, range=[${t.min},${t.max}], valid=${n>=t.min&&n<=t.max}`),n>=t.min&&n<=t.max&&t.setSize(n)}else console.debug(`[zoom] no saved value for ${t.storageKey}, using default=${t.default}`);t.el.addEventListener("mouseenter",()=>{Me=t}),t.el.addEventListener("mouseleave",()=>{if(Me===t){Me=null;for(const n of ra)n!==t&&n.el.contains(t.el)&&(!Me||Me.el.contains(n.el))&&(Me=n)}})}function qi(t,e){if(t.passthrough)return;const n=t.step??1,r=t.getSize()+e*n,i=Math.min(t.max,Math.max(t.min,Math.round(r*1e3)/1e3));t.setSize(i),localStorage.setItem(t.storageKey,String(i)),Qh(t,i)}function Jy(t){t.passthrough||(t.setSize(t.default),localStorage.setItem(t.storageKey,String(t.default)),Qh(t,t.default))}function Qh(t,e){if(t.group)for(const n of ra){if(n===t||n.group!==t.group||n.passthrough)continue;const s=Math.min(n.max,Math.max(n.min,e));n.setSize(s),localStorage.setItem(n.storageKey,String(s))}}let Kc=!1;function ev(){Kc||(Kc=!0,document.addEventListener("wheel",t=>{if(!t.ctrlKey||!Me||(t.preventDefault(),Me.passthrough))return;const e=t.deltaY<0?1:-1;qi(Me,e)},{passive:!1}),document.addEventListener("keydown",t=>{if(!(!t.ctrlKey||t.altKey||!Me)){if(t.key==="="||t.key==="+"){t.preventDefault(),qi(Me,1);return}if(t.key==="-"){t.preventDefault(),qi(Me,-1);return}t.key==="0"&&(t.preventDefault(),Jy(Me))}}))}function tv(t,e,n=13,s){const r=document.querySelector(t);return r?(qh({el:r,getSize:()=>parseFloat(r.style.zoom||"1"),setSize:i=>{r.style.zoom=String(i)},min:(s==null?void 0:s.min)??.6,max:(s==null?void 0:s.max)??2,default:1,step:.05,storageKey:e,passthrough:s==null?void 0:s.passthrough,group:s==null?void 0:s.group}),!0):!1}function nv(t,e,n){const s=document.querySelector(t);if(!s)return!1;const r=s,i=(n==null?void 0:n.defaultSize)??13,o=n==null?void 0:n.target;function a(){return o?Array.from(r.querySelectorAll(o)):[r]}return qh({el:s,getSize:()=>{const l=a();return l.length===0?i:parseFloat(l[0].style.fontSize||String(i))},setSize:l=>{const c=`${l}px`;for(const d of a())d.style.fontSize=c},min:(n==null?void 0:n.min)??8,max:(n==null?void 0:n.max)??24,default:i,step:(n==null?void 0:n.step)??1,storageKey:e,group:n==null?void 0:n.group}),!0}let lt=null,Vc=0;function sv(t,e,n){lt||(lt=document.createElement("div"),Object.assign(lt.style,{position:"fixed",padding:"2px 8px",borderRadius:"4px",background:"rgba(0,0,0,0.7)",color:"#fff",fontSize:"12px",fontFamily:"monospace",pointerEvents:"none",zIndex:"99999",opacity:"0",transition:"opacity 0.15s"}),document.body.appendChild(lt)),lt.textContent=t,lt.style.left=`${e}px`,lt.style.top=`${n}px`,lt.style.opacity="1",clearTimeout(Vc),Vc=window.setTimeout(()=>{lt&&(lt.style.opacity="0")},800)}function rv(){const t=Zy();if(!t||t.passthrough)return null;const e=t.getSize();return t.step&&t.step<1?`${Math.round(e*100)}%`:`${Math.round(e)}px`}let Xc=!1;function iv(){if(Xc)return;Xc=!0;let t=0,e=0;document.addEventListener("mousemove",s=>{t=s.clientX,e=s.clientY});const n=()=>requestAnimationFrame(()=>{const s=rv();s&&sv(s,t+16,e-24)});document.addEventListener("wheel",s=>{s.ctrlKey&&n()},{passive:!0}),document.addEventListener("keydown",s=>{s.ctrlKey&&!s.altKey&&"=+-0".includes(s.key)&&n()})}function ov(t=[],e=[],n=3e4){ev(),iv();const s=new Set,r=t.length+e.length;function i(){let a=0;for(const l of t){if(s.has(l.selector))continue;!!document.querySelector(l.selector)&&tv(l.selector,l.storageKey,1,l)&&(s.add(l.selector),a++,console.log("[zoom] CSS-zoom registered:",l.selector))}for(const l of e){if(s.has(l.selector))continue;!!document.querySelector(l.selector)&&nv(l.selector,l.storageKey,l)&&(s.add(l.selector),a++,console.log("[zoom] font-size registered:",l.selector))}return a}const o=i();if(console.log("[zoom] initial registration:",o,"/",r,"zones"),s.size<r){console.log("[zoom] starting MutationObserver for remaining",r-s.size,"zones");const a=new MutationObserver(()=>{const l=i();l>0&&console.log("[zoom] observer registered",l,"more zones, total:",s.size,"/",r),s.size>=r&&(a.disconnect(),console.log("[zoom] all zones registered, observer disconnected"))});a.observe(document.body,{childList:!0,subtree:!0}),setTimeout(()=>{a.disconnect(),s.size<r&&console.warn("[zoom] timeout: only",s.size,"/",r,"zones registered")},n)}}const av={orientation:"horizontal",cursor:"col-resize",pointerPos:t=>("clientX"in t,t.clientX),size:t=>t.offsetWidth,setSize:(t,e)=>{t.style.width=`${e}px`},clearSize:t=>{t.style.width=""},sizeProp:"width",maxSizeProp:"maxWidth",lockFlex:t=>{t.style.flexShrink="0",t.style.flexGrow="0"},unlockFlex:t=>{t.style.flexShrink="",t.style.flexGrow=""},chevrons:{beforeCollapsed:"fa-chevron-right",beforeExpanded:"fa-chevron-left",afterCollapsed:"fa-chevron-left",afterExpanded:"fa-chevron-right"}},lv={orientation:"vertical",cursor:"row-resize",pointerPos:t=>("clientY"in t,t.clientY),size:t=>t.offsetHeight,setSize:(t,e)=>{t.style.setProperty("height",`${e}px`,"important"),t.style.setProperty("flex",`0 0 ${e}px`,"important")},clearSize:t=>{t.style.height="",t.style.removeProperty("flex")},sizeProp:"height",maxSizeProp:"maxHeight",lockFlex:t=>{t.style.setProperty("flex-shrink","0","important"),t.style.setProperty("flex-grow","0","important")},unlockFlex:t=>{t.style.removeProperty("flex-shrink"),t.style.removeProperty("flex-grow"),t.style.removeProperty("flex")},chevrons:{beforeCollapsed:"fa-chevron-down",beforeExpanded:"fa-chevron-up",afterCollapsed:"fa-chevron-up",afterExpanded:"fa-chevron-down"}};function Yc(t){const n=getComputedStyle(t).flexDirection;return n==="column"||n==="column-reverse"?lv:av}function Gc(t,e,n){try{localStorage.setItem(t+e.storageKey,n.toString())}catch(s){console.warn("[WorkspacePanelResizer] Failed to save size:",s)}}function cv(t,e,n){try{if(e.fixedWidth)return;if(n.classList.contains("collapsed")){n.style.width="",n.style.height="",n.style.flexShrink="",n.style.flexGrow="",console.log(`[WorkspacePanelResizer] Panel ${e.storageKey} is collapsed, using CSS size`);return}const s=localStorage.getItem(t+e.storageKey);if(s){const r=parseInt(s,10);if(r>=e.minWidth){const i=n.closest(".workspace-three-col");i&&getComputedStyle(i).flexDirection==="column"?n.style.height=`${r}px`:n.style.width=`${r}px`,n.style.flexShrink="0",n.style.flexGrow="0",console.log(`[WorkspacePanelResizer] Restored ${e.storageKey} to ${r}px`)}}}catch(s){console.warn("[WorkspacePanelResizer] Failed to restore size:",s)}}function uv(t,e,n){const s=localStorage.getItem(t+e.storageKey);if(!s)return null;const r=parseInt(s,10);if(r<=e.minWidth+10)return null;const i=n.parentElement;if(i){const a=(getComputedStyle(i).flexDirection==="column"?i.offsetHeight:i.offsetWidth)*.8;if(r>a&&a>100)return null}return r}function qc(t,e,n,s){if(t.collapseStorageKey)try{const r=localStorage.getItem(t.collapseStorageKey);r==="true"?(e.classList.add("collapsed"),e.style.width="",e.style.height="",e.style.flexShrink="",e.style.flexGrow="",n&&s(n,t.resizeDirection,!0)):r==="false"&&(e.classList.remove("collapsed"),n&&s(n,t.resizeDirection,!1))}catch(r){console.warn("[WorkspacePanelResizer] Failed to restore collapse state:",r)}}function cn(t,e,n){const s=t.querySelector("i");s&&(e==="left"?n?(s.classList.remove("fa-chevron-left"),s.classList.add("fa-chevron-right")):(s.classList.remove("fa-chevron-right"),s.classList.add("fa-chevron-left")):n?(s.classList.remove("fa-chevron-right"),s.classList.add("fa-chevron-left")):(s.classList.remove("fa-chevron-left"),s.classList.add("fa-chevron-right")))}function Qc(t,e){if(!e.toggleButtonId)return;const n=document.getElementById(e.toggleButtonId),s=document.querySelector(e.targetPanel);if(!n||!s){console.warn(`[WorkspacePanelResizer] Missing toggle elements for ${e.toggleButtonId}`);return}if(n.dataset.wprToggleInit==="true"){console.log(`[WorkspacePanelResizer] Toggle already initialized for ${e.toggleButtonId}, skipping.`);return}n.dataset.wprToggleInit="true";const r=s.querySelector(".stx-shell-sidebar__header");r&&r.addEventListener("dblclick",()=>{n.click()}),n.addEventListener("click",i=>{i.preventDefault(),i.stopPropagation();const o=s.classList.toggle("collapsed");if(o)s.style.width="",s.style.height="",s.style.maxWidth="",s.style.maxHeight="",s.style.flexShrink="",s.style.flexGrow="";else{const a=uv(t,e,s);if(a){const l=s.closest(".workspace-three-col");l&&getComputedStyle(l).flexDirection==="column"?(s.style.height=`${a}px`,s.style.width=""):(s.style.width=`${a}px`,s.style.height=""),s.style.flexShrink="0",s.style.flexGrow="0"}else s.style.width="",s.style.height="",s.style.maxWidth="",s.style.maxHeight="",s.style.flexShrink="",s.style.flexGrow=""}cn(n,e.resizeDirection,o),e.collapseStorageKey&&localStorage.setItem(e.collapseStorageKey,o.toString())}),console.log(`[WorkspacePanelResizer] Toggle click handler attached for ${e.toggleButtonId}`)}const dv=32,hv=8;function Zc(t,e,n=dv,s=hv){let r=t,i=n;for(const c of e){const d=Math.abs(t-c);d<i&&(i=d,r=c)}if(r===t)return{value:t,snapped:!1};if(i<=s)return{value:r,snapped:!0};const o=1-(i-s)/(n-s),a=o*o;return{value:t+(r-t)*a,snapped:!0}}function Jc(t,e=[20,25,33,50,67,75,80]){return e.map(n=>Math.round(t*n/100))}const Qi=40,fv=40;function eu(t,e){const n=t.closest(".ws-ai-pane, .ws-worktree-pane, .ws-viewer-pane, .ws-apps-pane")||t.parentElement;if(!n)return 1/0;const s=n.closest(".workspace-three-col");if(!s)return 1/0;const r=e.size(s);let i=0;for(const o of Array.from(s.children))o!==n&&(o.classList.contains("ws-module-pane")?i+=fv:i+=e.size(o));return r-i}function Zi(t,e,n=!1){const s=document.querySelector(t.targetPanel);if(!s)return null;const r=s.closest(".ws-ai-pane, .ws-worktree-pane, .ws-viewer-pane, .ws-apps-pane, .ws-module-pane");if(!r)return null;let i=r;for(;i;){const o=e==="shrink-left"?i.previousElementSibling:i.nextElementSibling;if(!o)return null;const a=o.querySelector("[data-panel-resizer]"),l=o.querySelector(".stx-shell-sidebar");if(!a||!l){i=o;continue}if(l.classList.contains("collapsed")&&!n){i=o;continue}const c={resizerId:a.id,targetPanel:`#${l.id}`,minWidth:parseInt(a.dataset.minWidth||"40",10),storageKey:a.dataset.storageKey||"",resizeDirection:a.dataset.direction||"left",toggleButtonId:a.dataset.toggleBtn,collapseStorageKey:a.dataset.collapseKey,defaultWidth:a.dataset.defaultWidth?parseInt(a.dataset.defaultWidth,10):void 0};return{panel:l,config:c}}return null}function Ji(t,e,n,s){if(n.classList.add("collapsed"),s.clearSize(n),s.unlockFlex(n),e.toggleButtonId){const r=document.getElementById(e.toggleButtonId);r&&cn(r,e.resizeDirection,!0)}e.collapseStorageKey&&localStorage.setItem(e.collapseStorageKey,"true")}function pv(t,e){const n=document.getElementById(e.resizerId),s=document.querySelector(e.targetPanel);if(!n||!s){console.warn(`[WorkspacePanelResizer] Missing elements for ${e.resizerId}`);return}cv(t,e,s);let r=!1,i=0,o=0,a=!1,l=!1,c,d=s,u=null;const f=()=>{const S=n.closest(".workspace-three-col"),b=Yc(S||document.createElement("div"));return b.orientation==="vertical"?d=s.closest(".ws-ai-pane, .ws-worktree-pane, .ws-viewer-pane, .ws-apps-pane")||s:d=s,b},p=()=>{document.querySelectorAll(".stx-shell-sidebar").forEach(S=>{S.style.transition="none"})},g=()=>{document.querySelectorAll(".stx-shell-sidebar").forEach(S=>{S.style.transition=""})},w=S=>{if(c=f(),a=s.classList.contains("collapsed"),u=null,l=!1,e.fixedWidth){r=!0,i=c.pointerPos(S),o=c.size(d),l=!0,document.body.style.cursor=c.cursor,document.body.style.userSelect="none",n.classList.add("active"),p(),S.preventDefault();return}if(a){s.classList.remove("collapsed");const b=c.orientation==="vertical"?Math.max(150,c.size(d.parentElement||d)/4):e.minWidth;if(c.setSize(d,b),c.lockFlex(d),e.toggleButtonId){const T=document.getElementById(e.toggleButtonId);T&&cn(T,e.resizeDirection,!1)}e.collapseStorageKey&&localStorage.setItem(e.collapseStorageKey,"false")}r=!0,i=c.pointerPos(S),o=c.size(d),document.body.style.cursor=c.cursor,document.body.style.userSelect="none",n.classList.add("active"),d.dataset.wprDragging="true",p(),S.preventDefault()},E=S=>{if(!r)return;if(e.fixedWidth&&!u){const C=c.pointerPos(S)-i;if(Math.abs(C)<3)return;const F=e.resizeDirection==="left"?C<0:C>0,M=e.resizeDirection==="left"?"shrink-left":"shrink-right",N=Zi(e,M,!F);if(N){if(!F&&N.panel.classList.contains("collapsed")){if(N.panel.classList.remove("collapsed"),c.setSize(N.panel,N.config.minWidth),c.lockFlex(N.panel),N.config.toggleButtonId){const D=document.getElementById(N.config.toggleButtonId);D&&cn(D,N.config.resizeDirection,!1)}N.config.collapseStorageKey&&localStorage.setItem(N.config.collapseStorageKey,"false")}u={panel:N.panel,config:N.config,startSize:c.size(N.panel),startPos:c.pointerPos(S)}}return}if(l&&!u)return;if(u){const C=c.pointerPos(S)-u.startPos;let F=e.resizeDirection==="left"?u.startSize+C:u.startSize-C;const M=eu(u.panel,c);F>M&&(F=M);const N=u.panel.closest(".workspace-three-col");if(N){const D=Jc(c.size(N)),A=Zc(F,D);F=A.value,A.snapped?n.classList.add("snapped"):n.classList.remove("snapped")}if(F<Qi){Ji(t,u.config,u.panel,c);const D=e.resizeDirection==="left"?"shrink-left":"shrink-right",A=Zi(u.config,D);A?u={panel:A.panel,config:A.config,startSize:c.size(A.panel),startPos:c.pointerPos(S)}:u=null;return}c.setSize(u.panel,F),c.lockFlex(u.panel);return}const b=c.pointerPos(S)-i;let T=e.resizeDirection==="left"?o+b:o-b;const P=eu(d,c);T>P&&(T=P);const L=d.closest(".workspace-three-col");if(L){const C=Jc(c.size(L)),F=Zc(T,C);T=F.value,F.snapped?n.classList.add("snapped"):n.classList.remove("snapped")}if(T<Qi){l=!0,Ji(t,e,s,c);const C=e.resizeDirection==="left"?"shrink-left":"shrink-right",F=Zi(e,C);F&&(u={panel:F.panel,config:F.config,startSize:c.size(F.panel),startPos:c.pointerPos(S)});return}c.setSize(d,T),c.lockFlex(d)},m=()=>{if(!r)return;if(r=!1,document.body.style.cursor="",document.body.style.userSelect="",n.classList.remove("active"),n.classList.remove("snapped"),g(),requestAnimationFrame(()=>{delete d.dataset.wprDragging}),u){const b=c.size(u.panel);b>Qi+10&&Gc(t,u.config,b),u=null}if(s.classList.contains("collapsed")){a=!1;return}const S=c.size(d);S<=e.minWidth+10?Ji(t,e,s,c):Gc(t,e,S),a=!1},y=S=>{if(S.touches.length!==1)return;const b=S.touches[0];w({clientX:b.clientX,clientY:b.clientY,preventDefault:()=>S.preventDefault()})},v=S=>{if(!r||S.touches.length<1)return;S.preventDefault();const b=S.touches[0];E({clientX:b.clientX,clientY:b.clientY})},k=()=>{m()};n.addEventListener("mousedown",w),document.addEventListener("mousemove",E),document.addEventListener("mouseup",m),n.addEventListener("touchstart",y,{passive:!1}),document.addEventListener("touchmove",v,{passive:!1}),document.addEventListener("touchend",k,{passive:!0}),console.log(`[WorkspacePanelResizer] Initialized ${e.resizerId} (direction: ${e.resizeDirection}, touch: enabled)`)}class Zh{constructor(e="scitex-panel-"){this.panels=new Map,this.storagePrefix=e}initResizer(e){this.panels.set(e.resizerId,e),pv(this.storagePrefix,e)}initToggle(e){if(!e.toggleButtonId)return;const n=document.getElementById(e.toggleButtonId),s=document.querySelector(e.targetPanel);if(!n||!s){console.warn(`[WorkspacePanelResizer] Missing toggle elements for ${e.toggleButtonId}`);return}qc(e,s,n,cn),Qc(this.storagePrefix,e)}updateToggleIcon(e,n,s){cn(e,n,s)}initPanel(e){const n=document.querySelector(e.targetPanel),s=e.toggleButtonId?document.getElementById(e.toggleButtonId):null;n&&(n.style.transition="none",n.offsetWidth),n&&e.collapseStorageKey&&qc(e,n,s,cn),this.initResizer(e),Qc(this.storagePrefix,e),n&&requestAnimationFrame(()=>{requestAnimationFrame(()=>{n.style.transition=""})})}}let tu=!1;const ia=new Set;function Jh(t){const e={resizerId:t.id,targetPanel:t.dataset.target||"",minWidth:parseInt(t.dataset.minWidth||"40",10),storageKey:t.dataset.storageKey||"panel-width",resizeDirection:t.dataset.direction||"left",toggleButtonId:t.dataset.toggleBtn,collapseStorageKey:t.dataset.collapseKey,defaultWidth:t.dataset.defaultWidth?parseInt(t.dataset.defaultWidth,10):void 0,fixedWidth:t.hasAttribute("data-fixed-width")};return e.targetPanel?e:(console.warn("[WorkspacePanelResizer] Missing data-target on",t),null)}function nu(){if(tu){console.log("[WorkspacePanelResizer] autoInitPanels already done, skipping duplicate call.");return}tu=!0;const t=document.querySelectorAll("[data-panel-resizer]");document.body.classList.add("no-transition"),t.forEach(n=>{const s=n,r=s.dataset.storagePrefix||"scitex-",i=new Zh(r),o=Jh(s);o&&(i.initPanel(o),ia.add(s.id))}),console.log(`[WorkspacePanelResizer] Auto-initialized ${t.length} panel(s)`);const e=document.querySelector(".ws-module-pane");e&&new ResizeObserver(r=>{for(const i of r)i.contentRect.width<80?e.classList.add("collapsed"):e.classList.remove("collapsed")}).observe(e),requestAnimationFrame(()=>{requestAnimationFrame(()=>{document.querySelectorAll(".workspace-three-col, .workspace-shell").forEach(n=>n.classList.add("panels-ready")),document.body.classList.remove("no-transition")})})}function mv(){const t=document.querySelectorAll("[data-panel-resizer]");let e=0;document.body.classList.add("no-transition"),t.forEach(n=>{const s=n;if(!s.id||ia.has(s.id))return;const r=s.dataset.storagePrefix||"scitex-",i=new Zh(r),o=Jh(s);o&&(i.initPanel(o),ia.add(s.id),e++)}),e>0&&console.log(`[WorkspacePanelResizer] Late-initialized ${e} new panel(s)`),requestAnimationFrame(()=>{requestAnimationFrame(()=>{document.querySelectorAll(".workspace-three-col, .workspace-shell").forEach(n=>n.classList.add("panels-ready")),document.body.classList.remove("no-transition")})})}typeof document<"u"&&(document.readyState==="loading"?document.addEventListener("DOMContentLoaded",nu):nu(),document.addEventListener("htmx:afterSettle",()=>{mv()}));function gv(t,e){let n=t;for(const{prefix:s,handler:r}of e)n=vv(n,s,r);return n||null}function yv(t,e){const n=[];return n.push({prefix:"\x1B]9999;speak:",handler:s=>{try{const r=atob(s);if(!e){if("speechSynthesis"in window){const i=new SpeechSynthesisUtterance(r);speechSynthesis.speak(i)}}}catch{}}}),n.push({prefix:"\x1B]9998;media:",handler:s=>{try{const r=JSON.parse(atob(s));wv(r,t)}catch{}}}),n}function vv(t,e,n){const s=t.indexOf(e);if(s===-1)return t;const r=s+e.length,i=t.indexOf("\x07",r);return i===-1?t:(n(t.slice(r,i)),t.slice(0,s)+t.slice(i+1))}function wv(t,e){var o,a;const n=(a=(o=e.terminal)==null?void 0:o.element)==null?void 0:a.parentElement;if(!n)return;const s=document.createElement("div");s.className="stx-shell-terminal-media-overlay",s.style.cssText="position:absolute;top:0;left:0;right:0;bottom:0;background:rgba(0,0,0,0.85);display:flex;align-items:center;justify-content:center;z-index:100;";const r=document.createElement("button");r.className="stx-shell-terminal-media-close",r.innerHTML="&times;",r.style.cssText="position:absolute;top:8px;right:8px;background:none;border:none;color:#fff;font-size:24px;cursor:pointer;",r.onclick=()=>s.remove(),s.appendChild(r);const i=t.url||t.path;if(t.type==="image"){const l=document.createElement("img");l.src=i,l.style.maxWidth="100%",l.style.maxHeight="80%",s.appendChild(l)}else{const l=document.createElement("a");l.href=i,l.target="_blank",l.textContent=t.path.split("/").pop()||t.path,l.style.color="var(--color-accent-fg, #58a6ff)",s.appendChild(l)}n.style.position="relative",n.appendChild(s),setTimeout(()=>s.remove(),15e3)}const xv="/static/scitex_ui/vendor/xterm/xterm.js",Sv="/static/scitex_ui/vendor/xterm/xterm.css",Ev="/static/scitex_ui/vendor/xterm/xterm-addon-fit.js",kv="https://cdn.jsdelivr.net/npm/xterm@5.3.0/lib/xterm.js",bv="https://cdn.jsdelivr.net/npm/xterm@5.3.0/css/xterm.css",Cv="https://cdn.jsdelivr.net/npm/xterm-addon-fit@0.8.0/lib/xterm-addon-fit.js";let ur=null,eo=null;async function su(t,e){try{const n=await fetch(t);if(n.ok)return n.text()}catch{}return fetch(e).then(n=>n.text())}async function Mv(){var i,o;if(ur)return{Terminal:ur,FitAddon:eo};const t=window,[e,n]=await Promise.all([su(xv,kv),su(Ev,Cv)]),s=t.define,r=t.require;t.define=void 0,t.require=void 0;try{new Function(e)(),new Function(n)()}finally{t.define=s,t.require=r}return ur=((i=t.Terminal)==null?void 0:i.Terminal)||t.Terminal,eo=((o=t.FitAddon)==null?void 0:o.FitAddon)||t.FitAddon,{Terminal:ur,FitAddon:eo}}function Tv(){if(document.querySelector('link[href*="xterm.css"]'))return;const t=document.createElement("link");t.rel="stylesheet",t.href=Sv,t.onerror=()=>{t.href=bv},document.head.appendChild(t)}function Pv(){const t=getComputedStyle(document.documentElement),e=(s,r)=>t.getPropertyValue(s).trim()||r;return document.documentElement.getAttribute("data-theme")!=="light"?{background:e("--terminal-bg","#0d1117"),foreground:e("--terminal-fg","#c9d1d9"),cursor:e("--terminal-cursor","#58a6ff")}:{background:e("--terminal-bg","#ffffff"),foreground:e("--terminal-fg","#24292f"),cursor:e("--terminal-cursor","#0969da")}}function Lv(t,e,n){const s=new e({cursorBlink:!0,fontSize:13,fontFamily:"'JetBrains Mono', 'Monaco', 'Menlo', monospace",theme:Pv(),scrollback:1e4});s.open(t);const r={terminal:s,fitAddon:null,ws:null,connected:!1,resizeObserver:null,resizeTimeout:null};if(n){r.fitAddon=new n,s.loadAddon(r.fitAddon),ru(r);let i=0,o=0;r.resizeObserver=new ResizeObserver(a=>{const{width:l,height:c}=a[0].contentRect;Math.abs(l-i)<2&&Math.abs(c-o)<2||(i=l,o=c,r.resizeTimeout&&clearTimeout(r.resizeTimeout),r.resizeTimeout=setTimeout(()=>ru(r),150))}),r.resizeObserver.observe(t)}return s.onData(i=>{var o;((o=r.ws)==null?void 0:o.readyState)===WebSocket.OPEN&&r.ws.send(i)}),r}function ef(t){var e,n,s;((e=t.ws)==null?void 0:e.readyState)===WebSocket.OPEN&&((n=t.terminal)!=null&&n.rows)&&((s=t.terminal)!=null&&s.cols)&&t.ws.send(`resize:${t.terminal.rows}:${t.terminal.cols}`)}function tf(t,e,n,s=3e3){var i;if(t.connected)return;(i=e.onStatusChange)==null||i.call(e,"connecting");const r=e.getWebSocketUrl();console.log(`[TerminalFactory] Connecting to ${r}`),t.ws=new WebSocket(r),t.ws.onopen=()=>{var o;t.connected=!0,(o=e.onStatusChange)==null||o.call(e,"connected"),ef(t),console.log("[TerminalFactory] Connected")},t.ws.onmessage=o=>{let a=o.data;if(e.onMessage){const c=e.onMessage(a,t);if(c===null)return;a=c}const l=gv(a,n);l&&t.terminal.write(l)},t.ws.onerror=()=>{var o;(o=e.onStatusChange)==null||o.call(e,"error"),console.warn("[TerminalFactory] WebSocket error")},t.ws.onclose=()=>{var o;t.connected=!1,(o=e.onStatusChange)==null||o.call(e,"disconnected"),console.log(`[TerminalFactory] Disconnected, reconnecting in ${s}ms`),setTimeout(()=>tf(t,e,n,s),s)}}function ru(t){if(t.fitAddon)try{t.fitAddon.fit(),ef(t)}catch{}}function Fv(t){t.terminal.attachCustomKeyEventHandler(e=>{if(e.ctrlKey&&(e.key==="c"||e.key==="C")){const n=t.terminal.getSelection();return n?(navigator.clipboard.writeText(n),!1):!0}return e.ctrlKey&&(e.key==="v"||e.key==="V")?(navigator.clipboard.readText().then(n=>{var s;((s=t.ws)==null?void 0:s.readyState)===WebSocket.OPEN&&t.ws.send(n)}),!1):!0})}function Nv(t,e,n){t.addEventListener("dragover",s=>{s.preventDefault(),s.stopPropagation(),s.dataTransfer&&(s.dataTransfer.dropEffect="copy"),t.classList.add("stx-shell-drop-target")}),t.addEventListener("dragleave",()=>{t.classList.remove("stx-shell-drop-target")}),t.addEventListener("drop",s=>{var a;s.preventDefault(),s.stopPropagation(),t.classList.remove("stx-shell-drop-target");const r=s.dataTransfer;if(!r)return;if(r.files&&r.files.length>0&&n){(async()=>{var l;try{const c=await n(r.files);((l=e.ws)==null?void 0:l.readyState)===WebSocket.OPEN&&e.ws.send(c.join(" "))}catch(c){console.error("[TerminalFactory] File drop error:",c)}})();return}const o=(r.getData("text/plain")??"").split(";").filter(Boolean);o.length>0&&((a=e.ws)==null?void 0:a.readyState)===WebSocket.OPEN&&e.ws.send(o.join(" "))})}const _v="#stx-shell-ai-console-terminal";async function Iv(t){const e=typeof t.container=="string"?document.getElementById((t.container||_v).replace("#","")):t.container;if(!e)return console.log("[TerminalFactory] No terminal container found, skipping"),null;Tv();const{Terminal:n,FitAddon:s}=await Mv();if(!n)return console.error("[TerminalFactory] Failed to load xterm.js"),null;const r=Lv(e,n,s);Fv(r),t.dragDrop&&t.onFileDrop&&Nv(e,r,t.onFileDrop);const i=yv(r);return tf(r,t.adapter,i,t.reconnectDelay??3e3),console.log("[TerminalFactory] Initialized"),r}const iu="scitex_workspace_tree_",to=()=>({vis:null,writer:null,scholar:null,clew:null,hub:null,files:null,tools:null,explorer:null,all:null,code:null,example:null,apps:null});class Dv{constructor(e,n,s="all"){this.listeners=new Set,this.mode=s,this.projectKey=`${iu}${e}_${n}_${s}`,this.sharedKey=`${iu}${e}_${n}_shared`,this.state=this.loadState(),this.setupStorageListener()}loadState(){let e=new Set,n=null,s=new Set,r=new Set,i=0,o=to();try{const a=localStorage.getItem(this.sharedKey);if(a){const l=JSON.parse(a);e=new Set(l.expandedPaths||[])}}catch(a){console.warn("[TreeState] Failed to load shared state:",a)}try{const a=localStorage.getItem(this.projectKey);if(a){const l=JSON.parse(a);n=l.selectedPath||null,s=new Set(l.selectedPaths||[]),r=new Set(l.targetPaths||[]),i=l.scrollTop||0,o=l.focusPathPerMode||o}}catch(a){console.warn("[TreeState] Failed to load mode-specific state:",a)}return{expandedPaths:e,selectedPath:n,selectedPaths:s,targetPaths:r,scrollTop:i,focusPathPerMode:o,lastClickedPath:null}}saveState(){try{localStorage.setItem(this.sharedKey,JSON.stringify({expandedPaths:Array.from(this.state.expandedPaths)})),localStorage.setItem(this.projectKey,JSON.stringify({selectedPath:this.state.selectedPath,selectedPaths:Array.from(this.state.selectedPaths),targetPaths:Array.from(this.state.targetPaths),scrollTop:this.state.scrollTop,focusPathPerMode:this.state.focusPathPerMode}))}catch(e){console.warn("[TreeState] Failed to save state:",e)}}setupStorageListener(){window.addEventListener("storage",e=>{try{if(e.key===this.sharedKey&&e.newValue){const n=JSON.parse(e.newValue);this.state.expandedPaths=new Set(n.expandedPaths||[]),this.notifyListeners()}else if(e.key===this.projectKey&&e.newValue){const n=JSON.parse(e.newValue);this.state.selectedPath=n.selectedPath,this.state.selectedPaths=new Set(n.selectedPaths||[]),this.state.targetPaths=new Set(n.targetPaths||[]),this.state.scrollTop=n.scrollTop||0,this.state.focusPathPerMode=n.focusPathPerMode||to(),this.notifyListeners()}}catch(n){console.warn("[TreeState] Failed to parse storage event:",n)}})}notifyListeners(){this.listeners.forEach(e=>e(this.state))}subscribe(e){return this.listeners.add(e),()=>this.listeners.delete(e)}getState(){return this.state}isExpanded(e){return this.state.expandedPaths.has(e)}expand(e){this.state.expandedPaths.add(e),this.saveState(),this.notifyListeners()}collapse(e){this.state.expandedPaths.delete(e),this.saveState(),this.notifyListeners()}toggle(e){const n=this.isExpanded(e);return n?this.collapse(e):this.expand(e),!n}getExpanded(){return new Set(this.state.expandedPaths)}setSelected(e){this.state.selectedPath=e,this.saveState(),this.notifyListeners()}getSelected(){return this.state.selectedPath}isSelected(e){return this.state.selectedPaths.has(e)}addToSelection(e){this.state.selectedPaths.add(e),this.state.lastClickedPath=e,this.saveState(),this.notifyListeners()}removeFromSelection(e){this.state.selectedPaths.delete(e),this.saveState(),this.notifyListeners()}toggleSelection(e){this.state.selectedPaths.has(e)?this.state.selectedPaths.delete(e):this.state.selectedPaths.add(e),this.state.lastClickedPath=e,this.saveState(),this.notifyListeners()}getSelectedPaths(){return new Set(this.state.selectedPaths)}setSelectedPaths(e){this.state.selectedPaths=new Set(e),this.saveState(),this.notifyListeners()}clearSelection(){this.state.selectedPaths.clear(),this.state.selectedPath=null,this.state.lastClickedPath=null,this.saveState(),this.notifyListeners()}getLastClickedPath(){return this.state.lastClickedPath}setLastClickedPath(e){this.state.lastClickedPath=e}selectSingle(e){this.state.selectedPaths.clear(),this.state.selectedPaths.add(e),this.state.selectedPath=e,this.state.lastClickedPath=e,this.saveState(),this.notifyListeners()}selectSingleSilent(e){this.state.selectedPaths.clear(),this.state.selectedPaths.add(e),this.state.selectedPath=e,this.state.lastClickedPath=e,this.saveState()}setSelectedPathsSilent(e){this.state.selectedPaths=new Set(e),this.saveState()}setSelectedSilent(e){this.state.selectedPath=e,this.saveState()}setScrollTop(e){this.state.scrollTop=e,this.saveState()}getScrollTop(){return this.state.scrollTop}expandToPath(e){const n=e.split("/");let s="";for(let r=0;r<n.length-1;r++)s=s?`${s}/${n[r]}`:n[r],this.state.expandedPaths.add(s);this.saveState(),this.notifyListeners()}isTarget(e){return this.state.targetPaths.has(e)}addTarget(e){this.state.targetPaths.add(e),this.saveState(),this.notifyListeners()}removeTarget(e){this.state.targetPaths.delete(e),this.saveState(),this.notifyListeners()}setTargets(e){this.state.targetPaths=new Set(e),this.saveState(),this.notifyListeners()}clearTargets(){this.state.targetPaths.clear(),this.saveState(),this.notifyListeners()}getTargets(){return new Set(this.state.targetPaths)}setFocusPath(e,n){this.state.focusPathPerMode[e]=n,this.saveState(),this.notifyListeners()}getFocusPath(e){return this.state.focusPathPerMode[e]}clear(){this.state={expandedPaths:new Set,selectedPath:null,selectedPaths:new Set,targetPaths:new Set,scrollTop:0,focusPathPerMode:to(),lastClickedPath:null},localStorage.removeItem(this.projectKey),localStorage.removeItem(this.sharedKey),this.notifyListeners()}}const dr={code:{allowedExtensions:[],hiddenPatterns:["__pycache__",".pyc","node_modules",".git/objects"]},vis:{allowedExtensions:[".png",".jpg",".jpeg",".svg",".pdf",".csv",".json",".xlsx",".tsv"],hiddenPatterns:["__pycache__","node_modules",".git",".venv"]},writer:{allowedExtensions:[".tex",".bib",".pdf",".png",".jpg",".jpeg",".svg",".eps"],hiddenPatterns:["__pycache__","node_modules",".git",".venv","build",".aux",".log",".out"]},scholar:{allowedExtensions:[".bib"],hiddenPatterns:["__pycache__","node_modules",".git",".venv","build"]},clew:{allowedExtensions:[],hiddenPatterns:["__pycache__","node_modules",".git",".venv"]},hub:{allowedExtensions:[],hiddenPatterns:["__pycache__","node_modules",".git",".venv"]},files:{allowedExtensions:[],hiddenPatterns:["__pycache__","node_modules",".git"]},tools:{allowedExtensions:[],hiddenPatterns:["__pycache__","node_modules",".git",".venv"]},example:{allowedExtensions:[],hiddenPatterns:["__pycache__","node_modules",".git",".venv"]},apps:{allowedExtensions:[],hiddenPatterns:["__pycache__","node_modules",".git",".venv"]},explorer:{allowedExtensions:[],hiddenPatterns:["__pycache__","node_modules",".git",".venv"]},all:{allowedExtensions:[],disabledExtensions:[],hiddenPatterns:[]}},$v={code:["scripts"],vis:[],writer:[],scholar:[],clew:[],hub:[],files:[],tools:[],example:[],apps:[],explorer:[],all:[]},Rv={scholar:["scitex/scholar"],vis:["scitex/vis"],writer:["scitex/_writer"],code:[],clew:["scitex/verify"],hub:[],files:[],tools:[],explorer:[],example:[],apps:[],all:[]},ou={scholar:["node_modules",".git","__pycache__",".venv","venv","build","dist"],vis:["node_modules",".git","__pycache__",".venv","venv"],writer:["node_modules",".git","__pycache__",".venv","venv","build","dist","ai","config","docs","requirements","scripts","tests","texts","archive","figures","latex_styles","tables","wordcounts","logs","output"],code:[".git"],clew:["node_modules",".git","__pycache__",".venv","venv"],hub:["node_modules",".git","__pycache__",".venv","venv"],files:[".git"],tools:["node_modules",".git","__pycache__",".venv","venv"],explorer:["node_modules",".git","__pycache__",".venv","venv"],example:["node_modules",".git","__pycache__",".venv","venv"],apps:["node_modules",".git","__pycache__",".venv","venv"],all:[]},au={scholar:[".bib"],vis:[".csv",".tsv",".json",".xml",".png",".jpg",".jpeg",".gif",".svg",".webp",".pdf",".figz",".pltz"],writer:[".tex",".bib",".cls",".sty",".png",".jpg",".jpeg",".pdf",".svg",".eps",".csv",".tsv"],code:"all",clew:"all",hub:"all",files:"all",tools:"all",explorer:"all",example:"all",apps:"all",all:"all"},no=[".gitkeep"];class Av{constructor(e,n){this.showHidden=!1,this.moduleFilterEnabled=!1;const s=au[e],r=s==="all"?[]:s,i=ou[e]||[],o=dr[e]||dr.all;this.config={mode:e,allowedExtensions:(n==null?void 0:n.allowedExtensions)??r,disabledExtensions:(n==null?void 0:n.disabledExtensions)??o.disabledExtensions??[],hiddenPatterns:(n==null?void 0:n.hiddenPatterns)??i}}isHidden(e){const{name:n,type:s}=e;return!this.showHidden&&n.startsWith(".")?!0:s==="file"&&no.includes(n)?!1:!!(s==="file"&&[".DS_Store","Thumbs.db"].includes(n)||s==="directory"&&(n.endsWith(".figz.d")||n.endsWith(".pltz.d")))}isInactive(e){if(!this.moduleFilterEnabled)return!1;const{name:n,path:s,type:r}=e;for(const i of this.config.hiddenPatterns)if(i.startsWith(".")&&n.endsWith(i)||n===i||s.includes(`/${i}/`)||s.includes(`/${i}`))return!0;return!this.isWithinAllowedDirectories(s)||r==="file"&&!this.isAllowed(e)}isWithinAllowedDirectories(e){const n=Rv[this.config.mode];if(n.length===0)return!0;const s=e.replace(/^\.\//,"");return n.some(i=>{const o=i.replace(/^\.\//,"");return s.startsWith(o)||s===o})?!0:n.some(i=>i.replace(/^\.\//,"").startsWith(s+"/"))}isAllowed(e){if(e.type==="directory"||no.includes(e.name)||this.config.allowedExtensions.length===0)return!0;const n=this.getExtension(e.name);return this.config.allowedExtensions.includes(n)}isDisabled(e){if(!this.moduleFilterEnabled||e.type==="directory"||no.includes(e.name))return!1;const n=this.getExtension(e.name);return!!(this.config.disabledExtensions.includes(n)||this.config.allowedExtensions.length>0&&!this.config.allowedExtensions.includes(n))}getExtension(e){const n=e.lastIndexOf(".");return n===-1?"":e.substring(n).toLowerCase()}filterTree(e){return e.filter(n=>!this.isHidden(n)).map(n=>n.type==="directory"&&n.children?{...n,children:this.filterTree(n.children)}:n)}getMode(){return this.config.mode}getConfig(){return this.config}setAllowedExtensions(e){this.config.allowedExtensions=e}setDisabledExtensions(e){this.config.disabledExtensions=e}setHiddenPatterns(e){this.config.hiddenPatterns=e}setShowHidden(e){this.showHidden=e}getShowHidden(){return this.showHidden}setModuleFilterEnabled(e){this.moduleFilterEnabled=e}getModuleFilterEnabled(){return this.moduleFilterEnabled}setMode(e){const n=au[e],s=n==="all"?[]:n,r=ou[e]||[],i=dr[e]||dr.all;this.config={mode:e,allowedExtensions:s,disabledExtensions:i.disabledExtensions??[],hiddenPatterns:r}}}function jv(t){const e=t.substring(t.lastIndexOf(".")).toLowerCase(),n=t.toLowerCase(),s={".py":{icon:"fab fa-python",color:"#3776ab"},".js":{icon:"fab fa-js",color:"#f7df1e"},".ts":{icon:"fab fa-js",color:"#3178c6"},".jsx":{icon:"fab fa-react",color:"#61dafb"},".tsx":{icon:"fab fa-react",color:"#61dafb"},".java":{icon:"fab fa-java",color:"#007396"},".go":{icon:"fas fa-code",color:"#00add8"},".rs":{icon:"fas fa-gear",color:"#ce422b"},".c":{icon:"fas fa-file-code",color:"#555555"},".cpp":{icon:"fas fa-file-code",color:"#f34b7d"},".html":{icon:"fab fa-html5",color:"#e34c26"},".css":{icon:"fab fa-css3-alt",color:"#264de4"},".scss":{icon:"fab fa-sass",color:"#cc6699"},".json":{icon:"fas fa-brackets-curly",color:"#ffa500"},".yaml":{icon:"fas fa-file-code",color:"#cb171e"},".yml":{icon:"fas fa-file-code",color:"#cb171e"},".toml":{icon:"fas fa-file-code",color:"#9c4121"},".xml":{icon:"fas fa-file-code",color:"#e37933"},".csv":{icon:"fas fa-table",color:"#217346"},".md":{icon:"fab fa-markdown",color:"#000000"},".txt":{icon:"fas fa-file-alt",color:"#777777"},".pdf":{icon:"fas fa-file-pdf",color:"#ff0000"},".tex":{icon:"fas fa-file-alt",color:"#3d6117"},".bib":{icon:"fas fa-book-open",color:"#cb4b16"},".sh":{icon:"fas fa-terminal",color:"#89e051"},".bash":{icon:"fas fa-terminal",color:"#89e051"},".r":{icon:"fab fa-r-project",color:"#276dc3"},".ipynb":{icon:"fas fa-book",color:"#ff6f00"},".sql":{icon:"fas fa-database",color:"#e38c00"},".jpg":{icon:"fas fa-image",color:"#00bfa5"},".jpeg":{icon:"fas fa-image",color:"#00bfa5"},".png":{icon:"fas fa-image",color:"#00bfa5"},".gif":{icon:"fas fa-image",color:"#00bfa5"},".svg":{icon:"fas fa-image",color:"#ffb300"},".webp":{icon:"fas fa-image",color:"#00bfa5"},".zip":{icon:"fas fa-file-archive",color:"#8b8b8b"},".tar":{icon:"fas fa-file-archive",color:"#8b8b8b"},".gz":{icon:"fas fa-file-archive",color:"#8b8b8b"},".log":{icon:"fas fa-file-lines",color:"#777777"},".env":{icon:"fas fa-cog",color:"#edb92e"}},r={"readme.md":{icon:"fas fa-book",color:"#00b0ff"},readme:{icon:"fas fa-book",color:"#00b0ff"},license:{icon:"fas fa-scroll",color:"#ffab00"},dockerfile:{icon:"fab fa-docker",color:"#2496ed"},makefile:{icon:"fas fa-cog",color:"#6d6d6d"},".gitignore":{icon:"fab fa-git-alt",color:"#f05032"},".gitattributes":{icon:"fab fa-git-alt",color:"#f05032"},"package.json":{icon:"fab fa-npm",color:"#cb3837"},"tsconfig.json":{icon:"fab fa-js",color:"#3178c6"},"requirements.txt":{icon:"fab fa-python",color:"#3776ab"},"setup.py":{icon:"fab fa-python",color:"#3776ab"}};if(r[n]){const{icon:i,color:o}=r[n];return`<i class="${i}" style="color: ${o}; width: 16px; text-align: center;"></i>`}if(s[e]){const{icon:i,color:o}=s[e];return`<i class="${i}" style="color: ${o}; width: 16px; text-align: center;"></i>`}return'<i class="fas fa-file" style="color: #777; width: 16px; text-align: center;"></i>'}function zv(){return'<i class="fas fa-folder" style="color: #dcb67a; width: 16px; text-align: center;"></i>'}class Hv{constructor(e,n,s){this.searchMatches=new Set,this.searchAncestors=new Set,this.searchActive=!1,this.sortMode="name",this.config=e,this.stateManager=n,this.filter=s}renderLoadingSkeleton(){return`
+`}),r}getFailedResources(){const e=[];return window.performance&&window.performance.getEntriesByType&&window.performance.getEntriesByType("resource").forEach(s=>{s.responseStatus&&s.responseStatus>=400&&e.push(s.name)}),e}async captureDebugSnapshot(){this.notificationManager.showCameraFlash();const e=await this.captureScreenshotBlob(),n=this.getConsoleLogs();if(!e&&!n){this.notificationManager.showNotification("Copy failed","error"),this.notificationManager.triggerCopyCallback();return}if(e)try{await navigator.clipboard.write([new ClipboardItem({"image/png":e})]),this.notificationManager.showNotification("Screenshot copied - paste now!","success",2500)}catch(s){this.originalConsole.error("[ConsoleCollector] Screenshot clipboard failed:",s)}if(n&&n!=="No console logs captured."){const s=e?3e3:0;await new Promise(r=>setTimeout(r,s));try{await navigator.clipboard.writeText(n),this.notificationManager.showNotification("Console logs copied - paste now!","success")}catch(r){this.originalConsole.error("[ConsoleCollector] Logs clipboard failed:",r)}}else e||this.notificationManager.showNotification("No logs to copy","error");this.notificationManager.triggerCopyCallback()}async captureScreenshotBlob(){try{const e=await navigator.mediaDevices.getDisplayMedia({video:{displaySurface:"browser"},preferCurrentTab:!0,selfBrowserSurface:"include",systemAudio:"exclude"}),n=document.createElement("video");n.srcObject=e,n.muted=!0,await new Promise((o,a)=>{n.onloadedmetadata=()=>{n.play().then(()=>o()).catch(a)},n.onerror=a,setTimeout(()=>a(new Error("Video load timeout")),3e3)}),await new Promise(o=>setTimeout(o,100));const s=document.createElement("canvas");s.width=n.videoWidth,s.height=n.videoHeight;const r=s.getContext("2d");return r?(r.drawImage(n,0,0),e.getTracks().forEach(o=>o.stop()),await new Promise(o=>{s.toBlob(a=>o(a),"image/png")})):(e.getTracks().forEach(o=>o.stop()),null)}catch(e){return e.name!=="NotAllowedError"&&this.originalConsole.error("[ConsoleCollector] Screenshot failed:",e),null}}clearLogs(){this.consoleLogs=[],this.originalConsole.log("[ConsoleCollector] Logs cleared")}}class Bc{constructor(){this.isActive=!1,this.notificationManager=new Yy,this.debugCollector=new Vy,this.overlayManager=new Uy,this.elementScanner=new sa(this.debugCollector,this.notificationManager),this.selectionManager=new Xy(this.elementScanner.getElementBoxMap(),this.debugCollector,this.notificationManager),this.selectionManager.setElementScanner(this.elementScanner),new Gy(this.notificationManager),this.consoleCollector=new qy(this.notificationManager),this.notificationManager.setOnCopyCallback(()=>{this.deactivate()}),this.init()}init(){document.addEventListener("keydown",e=>{const n=e.key.toLowerCase();if(!["Tab","Enter","ArrowUp","ArrowDown","ArrowLeft","ArrowRight"].includes(e.key)){if(e.ctrlKey&&e.shiftKey&&!e.altKey&&n==="i"){e.preventDefault(),e.stopPropagation(),console.log("[ElementInspector] Ctrl+Shift+I pressed - capturing debug snapshot"),this.consoleCollector.captureDebugSnapshot();return}if(e.ctrlKey&&e.altKey&&!e.shiftKey&&n==="i"){e.preventDefault(),this.startSelectionMode();return}if(e.ctrlKey&&!e.altKey&&!e.shiftKey&&n==="i"&&this.isActive){e.preventDefault(),console.log("[ElementInspector] Ctrl+I pressed - loading next batch"),this.elementScanner.loadNextBatch();return}if(e.altKey&&!e.shiftKey&&!e.ctrlKey&&n==="i"){e.preventDefault(),this.toggle();return}if(e.key==="Escape"){this.selectionManager.isActive()?(e.preventDefault(),this.selectionManager.cancelSelectionMode(),this.deactivate()):this.isActive&&(e.preventDefault(),this.deactivate());return}}}),console.log("[ElementInspector] Initialized"),console.log("  Alt+I: Toggle inspector overlay"),console.log("  Ctrl+I: Load next 512 elements (when active)"),console.log("  Ctrl+Alt+I: Rectangle selection mode"),console.log("  Ctrl+Shift+I: Debug snapshot (screenshot + console logs)"),console.log("  Scroll wheel: Cycle through overlapped elements (affects rect selection depth)"),console.log("  Right-click: Copy element debug info"),console.log("  Left-click: Pass through to underlying element"),console.log("  Escape: Deactivate inspector / Cancel selection")}toggle(){this.isActive?this.deactivate():this.activate()}activate(){console.log("[ElementInspector] Activating..."),this.isActive=!0;const e=this.overlayManager.createOverlay();this.elementScanner.scanElements(e),console.log("[ElementInspector] Active - Press Alt+I to deactivate")}deactivate(){console.log("[ElementInspector] Deactivating..."),this.isActive=!1,this.elementScanner.clearElementBoxMap(),this.overlayManager.removeOverlay()}refresh(){this.isActive&&(this.deactivate(),this.activate())}startSelectionMode(){this.isActive||this.activate(),this.selectionManager.startSelectionMode()}}const Uc=window.elementInspector,Qy=Uc instanceof Bc?Uc:new Bc;window.elementInspector=Qy;let Wc;window.addEventListener("resize",()=>{clearTimeout(Wc),Wc=window.setTimeout(()=>{var t;(t=window.elementInspector)!=null&&t.isActive&&window.elementInspector.refresh()},500)});const ra=[];let Me=null;function Zy(){return Me}function qh(t){ra.push(t);const e=localStorage.getItem(t.storageKey);if(e){const n=parseFloat(e);console.debug(`[zoom] restoring ${t.storageKey}: saved=${e}, parsed=${n}, range=[${t.min},${t.max}], valid=${n>=t.min&&n<=t.max}`),n>=t.min&&n<=t.max&&t.setSize(n)}else console.debug(`[zoom] no saved value for ${t.storageKey}, using default=${t.default}`);t.el.addEventListener("mouseenter",()=>{Me=t}),t.el.addEventListener("mouseleave",()=>{if(Me===t){Me=null;for(const n of ra)n!==t&&n.el.contains(t.el)&&(!Me||Me.el.contains(n.el))&&(Me=n)}})}function qi(t,e){if(t.passthrough)return;const n=t.step??1,r=t.getSize()+e*n,i=Math.min(t.max,Math.max(t.min,Math.round(r*1e3)/1e3));t.setSize(i),localStorage.setItem(t.storageKey,String(i)),Qh(t,i)}function Jy(t){t.passthrough||(t.setSize(t.default),localStorage.setItem(t.storageKey,String(t.default)),Qh(t,t.default))}function Qh(t,e){if(t.group)for(const n of ra){if(n===t||n.group!==t.group||n.passthrough)continue;const s=Math.min(n.max,Math.max(n.min,e));n.setSize(s),localStorage.setItem(n.storageKey,String(s))}}let Kc=!1;function ev(){Kc||(Kc=!0,document.addEventListener("wheel",t=>{if(!t.ctrlKey||!Me||(t.preventDefault(),Me.passthrough))return;const e=t.deltaY<0?1:-1;qi(Me,e)},{passive:!1}),document.addEventListener("keydown",t=>{if(!(!t.ctrlKey||t.altKey||!Me)){if(t.key==="="||t.key==="+"){t.preventDefault(),qi(Me,1);return}if(t.key==="-"){t.preventDefault(),qi(Me,-1);return}t.key==="0"&&(t.preventDefault(),Jy(Me))}}))}function tv(t,e,n=13,s){const r=document.querySelector(t);return r?(qh({el:r,getSize:()=>parseFloat(r.style.zoom||"1"),setSize:i=>{r.style.zoom=String(i)},min:(s==null?void 0:s.min)??.6,max:(s==null?void 0:s.max)??2,default:1,step:.05,storageKey:e,passthrough:s==null?void 0:s.passthrough,group:s==null?void 0:s.group}),!0):!1}function nv(t,e,n){const s=document.querySelector(t);if(!s)return!1;const r=s,i=(n==null?void 0:n.defaultSize)??13,o=n==null?void 0:n.target;function a(){return o?Array.from(r.querySelectorAll(o)):[r]}return qh({el:s,getSize:()=>{const l=a();return l.length===0?i:parseFloat(l[0].style.fontSize||String(i))},setSize:l=>{const c=`${l}px`;for(const d of a())d.style.fontSize=c},min:(n==null?void 0:n.min)??8,max:(n==null?void 0:n.max)??24,default:i,step:(n==null?void 0:n.step)??1,storageKey:e,group:n==null?void 0:n.group}),!0}let lt=null,Vc=0;function sv(t,e,n){lt||(lt=document.createElement("div"),Object.assign(lt.style,{position:"fixed",padding:"2px 8px",borderRadius:"4px",background:"rgba(0,0,0,0.7)",color:"#fff",fontSize:"12px",fontFamily:"monospace",pointerEvents:"none",zIndex:"99999",opacity:"0",transition:"opacity 0.15s"}),document.body.appendChild(lt)),lt.textContent=t,lt.style.left=`${e}px`,lt.style.top=`${n}px`,lt.style.opacity="1",clearTimeout(Vc),Vc=window.setTimeout(()=>{lt&&(lt.style.opacity="0")},800)}function rv(){const t=Zy();if(!t||t.passthrough)return null;const e=t.getSize();return t.step&&t.step<1?`${Math.round(e*100)}%`:`${Math.round(e)}px`}let Xc=!1;function iv(){if(Xc)return;Xc=!0;let t=0,e=0;document.addEventListener("mousemove",s=>{t=s.clientX,e=s.clientY});const n=()=>requestAnimationFrame(()=>{const s=rv();s&&sv(s,t+16,e-24)});document.addEventListener("wheel",s=>{s.ctrlKey&&n()},{passive:!0}),document.addEventListener("keydown",s=>{s.ctrlKey&&!s.altKey&&"=+-0".includes(s.key)&&n()})}function ov(t=[],e=[],n=3e4){ev(),iv();const s=new Set,r=t.length+e.length;function i(){let a=0;for(const l of t){if(s.has(l.selector))continue;!!document.querySelector(l.selector)&&tv(l.selector,l.storageKey,1,l)&&(s.add(l.selector),a++,console.log("[zoom] CSS-zoom registered:",l.selector))}for(const l of e){if(s.has(l.selector))continue;!!document.querySelector(l.selector)&&nv(l.selector,l.storageKey,l)&&(s.add(l.selector),a++,console.log("[zoom] font-size registered:",l.selector))}return a}const o=i();if(console.log("[zoom] initial registration:",o,"/",r,"zones"),s.size<r){console.log("[zoom] starting MutationObserver for remaining",r-s.size,"zones");const a=new MutationObserver(()=>{const l=i();l>0&&console.log("[zoom] observer registered",l,"more zones, total:",s.size,"/",r),s.size>=r&&(a.disconnect(),console.log("[zoom] all zones registered, observer disconnected"))});a.observe(document.body,{childList:!0,subtree:!0}),setTimeout(()=>{a.disconnect(),s.size<r&&console.warn("[zoom] timeout: only",s.size,"/",r,"zones registered")},n)}}const av={orientation:"horizontal",cursor:"col-resize",pointerPos:t=>("clientX"in t,t.clientX),size:t=>t.offsetWidth,setSize:(t,e)=>{t.style.width=`${e}px`},clearSize:t=>{t.style.width=""},sizeProp:"width",maxSizeProp:"maxWidth",lockFlex:t=>{t.style.flexShrink="0",t.style.flexGrow="0"},unlockFlex:t=>{t.style.flexShrink="",t.style.flexGrow=""},chevrons:{beforeCollapsed:"fa-chevron-right",beforeExpanded:"fa-chevron-left",afterCollapsed:"fa-chevron-left",afterExpanded:"fa-chevron-right"}},lv={orientation:"vertical",cursor:"row-resize",pointerPos:t=>("clientY"in t,t.clientY),size:t=>t.offsetHeight,setSize:(t,e)=>{t.style.setProperty("height",`${e}px`,"important"),t.style.setProperty("flex",`0 0 ${e}px`,"important")},clearSize:t=>{t.style.height="",t.style.removeProperty("flex")},sizeProp:"height",maxSizeProp:"maxHeight",lockFlex:t=>{t.style.setProperty("flex-shrink","0","important"),t.style.setProperty("flex-grow","0","important")},unlockFlex:t=>{t.style.removeProperty("flex-shrink"),t.style.removeProperty("flex-grow"),t.style.removeProperty("flex")},chevrons:{beforeCollapsed:"fa-chevron-down",beforeExpanded:"fa-chevron-up",afterCollapsed:"fa-chevron-up",afterExpanded:"fa-chevron-down"}};function Yc(t){const n=getComputedStyle(t).flexDirection;return n==="column"||n==="column-reverse"?lv:av}function Gc(t,e,n){try{localStorage.setItem(t+e.storageKey,n.toString())}catch(s){console.warn("[WorkspacePanelResizer] Failed to save size:",s)}}function cv(t,e,n){try{if(e.fixedWidth)return;if(n.classList.contains("collapsed")){n.style.width="",n.style.height="",n.style.flexShrink="",n.style.flexGrow="",console.log(`[WorkspacePanelResizer] Panel ${e.storageKey} is collapsed, using CSS size`);return}const s=localStorage.getItem(t+e.storageKey);if(s){const r=parseInt(s,10);if(r>=e.minWidth){const i=n.closest(".workspace-three-col");i&&getComputedStyle(i).flexDirection==="column"?n.style.height=`${r}px`:n.style.width=`${r}px`,n.style.flexShrink="0",n.style.flexGrow="0",console.log(`[WorkspacePanelResizer] Restored ${e.storageKey} to ${r}px`)}}}catch(s){console.warn("[WorkspacePanelResizer] Failed to restore size:",s)}}function uv(t,e,n){const s=localStorage.getItem(t+e.storageKey);if(!s)return null;const r=parseInt(s,10);if(r<=e.minWidth+10)return null;const i=n.parentElement;if(i){const a=(getComputedStyle(i).flexDirection==="column"?i.offsetHeight:i.offsetWidth)*.8;if(r>a&&a>100)return null}return r}function qc(t,e,n,s){if(t.collapseStorageKey)try{const r=localStorage.getItem(t.collapseStorageKey);r==="true"?(e.classList.add("collapsed"),e.style.width="",e.style.height="",e.style.flexShrink="",e.style.flexGrow="",n&&s(n,t.resizeDirection,!0)):r==="false"&&(e.classList.remove("collapsed"),n&&s(n,t.resizeDirection,!1))}catch(r){console.warn("[WorkspacePanelResizer] Failed to restore collapse state:",r)}}function cn(t,e,n){const s=t.querySelector("i");s&&(e==="left"?n?(s.classList.remove("fa-chevron-left"),s.classList.add("fa-chevron-right")):(s.classList.remove("fa-chevron-right"),s.classList.add("fa-chevron-left")):n?(s.classList.remove("fa-chevron-right"),s.classList.add("fa-chevron-left")):(s.classList.remove("fa-chevron-left"),s.classList.add("fa-chevron-right")))}function Qc(t,e){if(!e.toggleButtonId)return;const n=document.getElementById(e.toggleButtonId),s=document.querySelector(e.targetPanel);if(!n||!s){console.warn(`[WorkspacePanelResizer] Missing toggle elements for ${e.toggleButtonId}`);return}if(n.dataset.wprToggleInit==="true"){console.log(`[WorkspacePanelResizer] Toggle already initialized for ${e.toggleButtonId}, skipping.`);return}n.dataset.wprToggleInit="true";const r=s.querySelector(".stx-shell-sidebar__header");r&&r.addEventListener("dblclick",()=>{n.click()}),n.addEventListener("click",i=>{i.preventDefault(),i.stopPropagation();const o=s.classList.toggle("collapsed");if(o){s.style.width="",s.style.height="",s.style.maxWidth="",s.style.maxHeight="",s.style.flexShrink="",s.style.flexGrow="";const a=s.closest(".ws-ai-pane, .ws-worktree-pane, .ws-viewer-pane, .ws-apps-pane");a&&(a.style.flex="",a.style.height="",a.style.minHeight="",a.style.maxHeight="")}else{const a=uv(t,e,s);if(a){const l=s.closest(".workspace-three-col");l&&getComputedStyle(l).flexDirection==="column"?(s.style.height=`${a}px`,s.style.width=""):(s.style.width=`${a}px`,s.style.height=""),s.style.flexShrink="0",s.style.flexGrow="0"}else s.style.width="",s.style.height="",s.style.maxWidth="",s.style.maxHeight="",s.style.flexShrink="",s.style.flexGrow=""}if(!o){const a=s.closest(".ws-ai-pane, .ws-worktree-pane, .ws-viewer-pane, .ws-apps-pane");a&&(a.style.flex="",a.style.height="",a.style.minHeight="",a.style.maxHeight="")}cn(n,e.resizeDirection,o),e.collapseStorageKey&&localStorage.setItem(e.collapseStorageKey,o.toString())}),console.log(`[WorkspacePanelResizer] Toggle click handler attached for ${e.toggleButtonId}`)}const dv=32,hv=8;function Zc(t,e,n=dv,s=hv){let r=t,i=n;for(const c of e){const d=Math.abs(t-c);d<i&&(i=d,r=c)}if(r===t)return{value:t,snapped:!1};if(i<=s)return{value:r,snapped:!0};const o=1-(i-s)/(n-s),a=o*o;return{value:t+(r-t)*a,snapped:!0}}function Jc(t,e=[20,25,33,50,67,75,80]){return e.map(n=>Math.round(t*n/100))}const Qi=40,fv=40;function eu(t,e){const n=t.closest(".ws-ai-pane, .ws-worktree-pane, .ws-viewer-pane, .ws-apps-pane")||t.parentElement;if(!n)return 1/0;const s=n.closest(".workspace-three-col");if(!s)return 1/0;const r=e.size(s);let i=0;for(const o of Array.from(s.children))o!==n&&(o.classList.contains("ws-module-pane")?i+=fv:i+=e.size(o));return r-i}function Zi(t,e,n=!1){const s=document.querySelector(t.targetPanel);if(!s)return null;const r=s.closest(".ws-ai-pane, .ws-worktree-pane, .ws-viewer-pane, .ws-apps-pane, .ws-module-pane");if(!r)return null;let i=r;for(;i;){const o=e==="shrink-left"?i.previousElementSibling:i.nextElementSibling;if(!o)return null;const a=o.querySelector("[data-panel-resizer]"),l=o.querySelector(".stx-shell-sidebar");if(!a||!l){i=o;continue}if(l.classList.contains("collapsed")&&!n){i=o;continue}const c={resizerId:a.id,targetPanel:`#${l.id}`,minWidth:parseInt(a.dataset.minWidth||"40",10),storageKey:a.dataset.storageKey||"",resizeDirection:a.dataset.direction||"left",toggleButtonId:a.dataset.toggleBtn,collapseStorageKey:a.dataset.collapseKey,defaultWidth:a.dataset.defaultWidth?parseInt(a.dataset.defaultWidth,10):void 0};return{panel:l,config:c}}return null}function Ji(t,e,n,s){if(n.classList.add("collapsed"),s.clearSize(n),s.unlockFlex(n),e.toggleButtonId){const r=document.getElementById(e.toggleButtonId);r&&cn(r,e.resizeDirection,!0)}e.collapseStorageKey&&localStorage.setItem(e.collapseStorageKey,"true")}function pv(t,e){const n=document.getElementById(e.resizerId),s=document.querySelector(e.targetPanel);if(!n||!s){console.warn(`[WorkspacePanelResizer] Missing elements for ${e.resizerId}`);return}cv(t,e,s);let r=!1,i=0,o=0,a=!1,l=!1,c,d=s,u=null;const f=()=>{const S=n.closest(".workspace-three-col"),b=Yc(S||document.createElement("div"));return b.orientation==="vertical"?d=s.closest(".ws-ai-pane, .ws-worktree-pane, .ws-viewer-pane, .ws-apps-pane")||s:d=s,b},p=()=>{document.querySelectorAll(".stx-shell-sidebar").forEach(S=>{S.style.transition="none"})},g=()=>{document.querySelectorAll(".stx-shell-sidebar").forEach(S=>{S.style.transition=""})},w=S=>{if(c=f(),a=s.classList.contains("collapsed"),u=null,l=!1,e.fixedWidth){r=!0,i=c.pointerPos(S),o=c.size(d),l=!0,document.body.style.cursor=c.cursor,document.body.style.userSelect="none",n.classList.add("active"),p(),S.preventDefault();return}if(a){s.classList.remove("collapsed");const b=c.orientation==="vertical"?Math.max(150,c.size(d.parentElement||d)/4):e.minWidth;if(c.setSize(d,b),c.lockFlex(d),e.toggleButtonId){const T=document.getElementById(e.toggleButtonId);T&&cn(T,e.resizeDirection,!1)}e.collapseStorageKey&&localStorage.setItem(e.collapseStorageKey,"false")}r=!0,i=c.pointerPos(S),o=c.size(d),document.body.style.cursor=c.cursor,document.body.style.userSelect="none",n.classList.add("active"),d.dataset.wprDragging="true",p(),S.preventDefault()},E=S=>{if(!r)return;if(e.fixedWidth&&!u){const C=c.pointerPos(S)-i;if(Math.abs(C)<3)return;const F=e.resizeDirection==="left"?C<0:C>0,M=e.resizeDirection==="left"?"shrink-left":"shrink-right",N=Zi(e,M,!F);if(N){if(!F&&N.panel.classList.contains("collapsed")){if(N.panel.classList.remove("collapsed"),c.setSize(N.panel,N.config.minWidth),c.lockFlex(N.panel),N.config.toggleButtonId){const D=document.getElementById(N.config.toggleButtonId);D&&cn(D,N.config.resizeDirection,!1)}N.config.collapseStorageKey&&localStorage.setItem(N.config.collapseStorageKey,"false")}u={panel:N.panel,config:N.config,startSize:c.size(N.panel),startPos:c.pointerPos(S)}}return}if(l&&!u)return;if(u){const C=c.pointerPos(S)-u.startPos;let F=e.resizeDirection==="left"?u.startSize+C:u.startSize-C;const M=eu(u.panel,c);F>M&&(F=M);const N=u.panel.closest(".workspace-three-col");if(N){const D=Jc(c.size(N)),A=Zc(F,D);F=A.value,A.snapped?n.classList.add("snapped"):n.classList.remove("snapped")}if(F<Qi){Ji(t,u.config,u.panel,c);const D=e.resizeDirection==="left"?"shrink-left":"shrink-right",A=Zi(u.config,D);A?u={panel:A.panel,config:A.config,startSize:c.size(A.panel),startPos:c.pointerPos(S)}:u=null;return}c.setSize(u.panel,F),c.lockFlex(u.panel);return}const b=c.pointerPos(S)-i;let T=e.resizeDirection==="left"?o+b:o-b;const P=eu(d,c);T>P&&(T=P);const L=d.closest(".workspace-three-col");if(L){const C=Jc(c.size(L)),F=Zc(T,C);T=F.value,F.snapped?n.classList.add("snapped"):n.classList.remove("snapped")}if(T<Qi){l=!0,Ji(t,e,s,c);const C=e.resizeDirection==="left"?"shrink-left":"shrink-right",F=Zi(e,C);F&&(u={panel:F.panel,config:F.config,startSize:c.size(F.panel),startPos:c.pointerPos(S)});return}c.setSize(d,T),c.lockFlex(d)},m=()=>{if(!r)return;if(r=!1,document.body.style.cursor="",document.body.style.userSelect="",n.classList.remove("active"),n.classList.remove("snapped"),g(),requestAnimationFrame(()=>{delete d.dataset.wprDragging}),u){const b=c.size(u.panel);b>Qi+10&&Gc(t,u.config,b),u=null}if(s.classList.contains("collapsed")){a=!1;return}const S=c.size(d);S<=e.minWidth+10?Ji(t,e,s,c):Gc(t,e,S),a=!1},y=S=>{if(S.touches.length!==1)return;const b=S.touches[0];w({clientX:b.clientX,clientY:b.clientY,preventDefault:()=>S.preventDefault()})},v=S=>{if(!r||S.touches.length<1)return;S.preventDefault();const b=S.touches[0];E({clientX:b.clientX,clientY:b.clientY})},k=()=>{m()};n.addEventListener("mousedown",w),document.addEventListener("mousemove",E),document.addEventListener("mouseup",m),n.addEventListener("touchstart",y,{passive:!1}),document.addEventListener("touchmove",v,{passive:!1}),document.addEventListener("touchend",k,{passive:!0}),console.log(`[WorkspacePanelResizer] Initialized ${e.resizerId} (direction: ${e.resizeDirection}, touch: enabled)`)}class Zh{constructor(e="scitex-panel-"){this.panels=new Map,this.storagePrefix=e}initResizer(e){this.panels.set(e.resizerId,e),pv(this.storagePrefix,e)}initToggle(e){if(!e.toggleButtonId)return;const n=document.getElementById(e.toggleButtonId),s=document.querySelector(e.targetPanel);if(!n||!s){console.warn(`[WorkspacePanelResizer] Missing toggle elements for ${e.toggleButtonId}`);return}qc(e,s,n,cn),Qc(this.storagePrefix,e)}updateToggleIcon(e,n,s){cn(e,n,s)}initPanel(e){const n=document.querySelector(e.targetPanel),s=e.toggleButtonId?document.getElementById(e.toggleButtonId):null;n&&(n.style.transition="none",n.offsetWidth),n&&e.collapseStorageKey&&qc(e,n,s,cn),this.initResizer(e),Qc(this.storagePrefix,e),n&&requestAnimationFrame(()=>{requestAnimationFrame(()=>{n.style.transition=""})})}}let tu=!1;const ia=new Set;function Jh(t){const e={resizerId:t.id,targetPanel:t.dataset.target||"",minWidth:parseInt(t.dataset.minWidth||"40",10),storageKey:t.dataset.storageKey||"panel-width",resizeDirection:t.dataset.direction||"left",toggleButtonId:t.dataset.toggleBtn,collapseStorageKey:t.dataset.collapseKey,defaultWidth:t.dataset.defaultWidth?parseInt(t.dataset.defaultWidth,10):void 0,fixedWidth:t.hasAttribute("data-fixed-width")};return e.targetPanel?e:(console.warn("[WorkspacePanelResizer] Missing data-target on",t),null)}function nu(){if(tu){console.log("[WorkspacePanelResizer] autoInitPanels already done, skipping duplicate call.");return}tu=!0;const t=document.querySelectorAll("[data-panel-resizer]");document.body.classList.add("no-transition"),t.forEach(n=>{const s=n,r=s.dataset.storagePrefix||"scitex-",i=new Zh(r),o=Jh(s);o&&(i.initPanel(o),ia.add(s.id))}),console.log(`[WorkspacePanelResizer] Auto-initialized ${t.length} panel(s)`);const e=document.querySelector(".ws-module-pane");e&&new ResizeObserver(r=>{for(const i of r)i.contentRect.width<80?e.classList.add("collapsed"):e.classList.remove("collapsed")}).observe(e),requestAnimationFrame(()=>{requestAnimationFrame(()=>{document.querySelectorAll(".workspace-three-col, .workspace-shell").forEach(n=>n.classList.add("panels-ready")),document.body.classList.remove("no-transition"),document.body.classList.add("app-ready")})})}function mv(){const t=document.querySelectorAll("[data-panel-resizer]");let e=0;document.body.classList.add("no-transition"),t.forEach(n=>{const s=n;if(!s.id||ia.has(s.id))return;const r=s.dataset.storagePrefix||"scitex-",i=new Zh(r),o=Jh(s);o&&(i.initPanel(o),ia.add(s.id),e++)}),e>0&&console.log(`[WorkspacePanelResizer] Late-initialized ${e} new panel(s)`),requestAnimationFrame(()=>{requestAnimationFrame(()=>{document.querySelectorAll(".workspace-three-col, .workspace-shell").forEach(n=>n.classList.add("panels-ready")),document.body.classList.remove("no-transition")})})}typeof document<"u"&&(document.readyState==="loading"?document.addEventListener("DOMContentLoaded",nu):nu(),document.addEventListener("htmx:afterSettle",()=>{mv()}),window.addEventListener("load",()=>{document.body.classList.contains("app-ready")||document.body.classList.add("app-ready")}));function gv(t,e){let n=t;for(const{prefix:s,handler:r}of e)n=vv(n,s,r);return n||null}function yv(t,e){const n=[];return n.push({prefix:"\x1B]9999;speak:",handler:s=>{try{const r=atob(s);if(!e){if("speechSynthesis"in window){const i=new SpeechSynthesisUtterance(r);speechSynthesis.speak(i)}}}catch{}}}),n.push({prefix:"\x1B]9998;media:",handler:s=>{try{const r=JSON.parse(atob(s));wv(r,t)}catch{}}}),n}function vv(t,e,n){const s=t.indexOf(e);if(s===-1)return t;const r=s+e.length,i=t.indexOf("\x07",r);return i===-1?t:(n(t.slice(r,i)),t.slice(0,s)+t.slice(i+1))}function wv(t,e){var o,a;const n=(a=(o=e.terminal)==null?void 0:o.element)==null?void 0:a.parentElement;if(!n)return;const s=document.createElement("div");s.className="stx-shell-terminal-media-overlay",s.style.cssText="position:absolute;top:0;left:0;right:0;bottom:0;background:rgba(0,0,0,0.85);display:flex;align-items:center;justify-content:center;z-index:100;";const r=document.createElement("button");r.className="stx-shell-terminal-media-close",r.innerHTML="&times;",r.style.cssText="position:absolute;top:8px;right:8px;background:none;border:none;color:#fff;font-size:24px;cursor:pointer;",r.onclick=()=>s.remove(),s.appendChild(r);const i=t.url||t.path;if(t.type==="image"){const l=document.createElement("img");l.src=i,l.style.maxWidth="100%",l.style.maxHeight="80%",s.appendChild(l)}else{const l=document.createElement("a");l.href=i,l.target="_blank",l.textContent=t.path.split("/").pop()||t.path,l.style.color="var(--color-accent-fg, #58a6ff)",s.appendChild(l)}n.style.position="relative",n.appendChild(s),setTimeout(()=>s.remove(),15e3)}const xv="/static/scitex_ui/vendor/xterm/xterm.js",Sv="/static/scitex_ui/vendor/xterm/xterm.css",Ev="/static/scitex_ui/vendor/xterm/xterm-addon-fit.js",kv="https://cdn.jsdelivr.net/npm/xterm@5.3.0/lib/xterm.js",bv="https://cdn.jsdelivr.net/npm/xterm@5.3.0/css/xterm.css",Cv="https://cdn.jsdelivr.net/npm/xterm-addon-fit@0.8.0/lib/xterm-addon-fit.js";let ur=null,eo=null;async function su(t,e){try{const n=await fetch(t);if(n.ok)return n.text()}catch{}return fetch(e).then(n=>n.text())}async function Mv(){var i,o;if(ur)return{Terminal:ur,FitAddon:eo};const t=window,[e,n]=await Promise.all([su(xv,kv),su(Ev,Cv)]),s=t.define,r=t.require;t.define=void 0,t.require=void 0;try{new Function(e)(),new Function(n)()}finally{t.define=s,t.require=r}return ur=((i=t.Terminal)==null?void 0:i.Terminal)||t.Terminal,eo=((o=t.FitAddon)==null?void 0:o.FitAddon)||t.FitAddon,{Terminal:ur,FitAddon:eo}}function Tv(){if(document.querySelector('link[href*="xterm.css"]'))return;const t=document.createElement("link");t.rel="stylesheet",t.href=Sv,t.onerror=()=>{t.href=bv},document.head.appendChild(t)}function Pv(){const t=getComputedStyle(document.documentElement),e=(s,r)=>t.getPropertyValue(s).trim()||r;return document.documentElement.getAttribute("data-theme")!=="light"?{background:e("--terminal-bg","#0d1117"),foreground:e("--terminal-fg","#c9d1d9"),cursor:e("--terminal-cursor","#58a6ff")}:{background:e("--terminal-bg","#ffffff"),foreground:e("--terminal-fg","#24292f"),cursor:e("--terminal-cursor","#0969da")}}function Lv(t,e,n){const s=new e({cursorBlink:!0,fontSize:13,fontFamily:"'JetBrains Mono', 'Monaco', 'Menlo', monospace",theme:Pv(),scrollback:1e5});s.open(t);const r={terminal:s,fitAddon:null,ws:null,connected:!1,resizeObserver:null,resizeTimeout:null};if(n){r.fitAddon=new n,s.loadAddon(r.fitAddon),ru(r);let i=0,o=0;r.resizeObserver=new ResizeObserver(a=>{const{width:l,height:c}=a[0].contentRect;Math.abs(l-i)<2&&Math.abs(c-o)<2||(i=l,o=c,r.resizeTimeout&&clearTimeout(r.resizeTimeout),r.resizeTimeout=setTimeout(()=>ru(r),150))}),r.resizeObserver.observe(t)}return s.onData(i=>{var o;((o=r.ws)==null?void 0:o.readyState)===WebSocket.OPEN&&r.ws.send(i)}),r}function ef(t){var e,n,s;((e=t.ws)==null?void 0:e.readyState)===WebSocket.OPEN&&((n=t.terminal)!=null&&n.rows)&&((s=t.terminal)!=null&&s.cols)&&t.ws.send(`resize:${t.terminal.rows}:${t.terminal.cols}`)}function tf(t,e,n,s=3e3){var a;if(t.connected)return;(a=e.onStatusChange)==null||a.call(e,"connecting");const r=e.getWebSocketUrl();console.log(`[TerminalFactory] Connecting to ${r}`),t.ws=new WebSocket(r),t.ws.onopen=()=>{var l;t.connected=!0,(l=e.onStatusChange)==null||l.call(e,"connected"),ef(t),console.log("[TerminalFactory] Connected")};let i="",o=0;t.ws.onmessage=l=>{let c=l.data;if(e.onMessage){const u=e.onMessage(c,t);if(u===null)return;c=u}const d=gv(c,n);d&&(i+=d,o||(o=requestAnimationFrame(()=>{i&&t.terminal.write(i),i="",o=0})))},t.ws.onerror=()=>{var l;(l=e.onStatusChange)==null||l.call(e,"error"),console.warn("[TerminalFactory] WebSocket error")},t.ws.onclose=()=>{var l;t.connected=!1,(l=e.onStatusChange)==null||l.call(e,"disconnected"),console.log(`[TerminalFactory] Disconnected, reconnecting in ${s}ms`),setTimeout(()=>tf(t,e,n,s),s)}}function ru(t){if(t.fitAddon)try{t.fitAddon.fit(),ef(t)}catch{}}function Fv(t){t.terminal.attachCustomKeyEventHandler(e=>{if(e.ctrlKey&&(e.key==="c"||e.key==="C")){const n=t.terminal.getSelection();return n?(navigator.clipboard.writeText(n),!1):!0}return e.ctrlKey&&(e.key==="v"||e.key==="V")?(navigator.clipboard.readText().then(n=>{var s;((s=t.ws)==null?void 0:s.readyState)===WebSocket.OPEN&&t.ws.send(n)}),!1):!0})}function Nv(t,e,n){t.addEventListener("dragover",s=>{s.preventDefault(),s.stopPropagation(),s.dataTransfer&&(s.dataTransfer.dropEffect="copy"),t.classList.add("stx-shell-drop-target")}),t.addEventListener("dragleave",()=>{t.classList.remove("stx-shell-drop-target")}),t.addEventListener("drop",s=>{var a;s.preventDefault(),s.stopPropagation(),t.classList.remove("stx-shell-drop-target");const r=s.dataTransfer;if(!r)return;if(r.files&&r.files.length>0&&n){(async()=>{var l;try{const c=await n(r.files);((l=e.ws)==null?void 0:l.readyState)===WebSocket.OPEN&&e.ws.send(c.join(" "))}catch(c){console.error("[TerminalFactory] File drop error:",c)}})();return}const o=(r.getData("text/plain")??"").split(";").filter(Boolean);o.length>0&&((a=e.ws)==null?void 0:a.readyState)===WebSocket.OPEN&&e.ws.send(o.join(" "))})}const _v="#stx-shell-ai-console-terminal";async function Iv(t){const e=typeof t.container=="string"?document.getElementById((t.container||_v).replace("#","")):t.container;if(!e)return console.log("[TerminalFactory] No terminal container found, skipping"),null;Tv();const{Terminal:n,FitAddon:s}=await Mv();if(!n)return console.error("[TerminalFactory] Failed to load xterm.js"),null;const r=Lv(e,n,s);Fv(r),t.dragDrop&&t.onFileDrop&&Nv(e,r,t.onFileDrop);const i=yv(r);return tf(r,t.adapter,i,t.reconnectDelay??3e3),console.log("[TerminalFactory] Initialized"),r}const iu="scitex_workspace_tree_",to=()=>({vis:null,writer:null,scholar:null,clew:null,hub:null,files:null,tools:null,explorer:null,all:null,code:null,example:null,apps:null});class Dv{constructor(e,n,s="all"){this.listeners=new Set,this.mode=s,this.projectKey=`${iu}${e}_${n}_${s}`,this.sharedKey=`${iu}${e}_${n}_shared`,this.state=this.loadState(),this.setupStorageListener()}loadState(){let e=new Set,n=null,s=new Set,r=new Set,i=0,o=to();try{const a=localStorage.getItem(this.sharedKey);if(a){const l=JSON.parse(a);e=new Set(l.expandedPaths||[])}}catch(a){console.warn("[TreeState] Failed to load shared state:",a)}try{const a=localStorage.getItem(this.projectKey);if(a){const l=JSON.parse(a);n=l.selectedPath||null,s=new Set(l.selectedPaths||[]),r=new Set(l.targetPaths||[]),i=l.scrollTop||0,o=l.focusPathPerMode||o}}catch(a){console.warn("[TreeState] Failed to load mode-specific state:",a)}return{expandedPaths:e,selectedPath:n,selectedPaths:s,targetPaths:r,scrollTop:i,focusPathPerMode:o,lastClickedPath:null}}saveState(){try{localStorage.setItem(this.sharedKey,JSON.stringify({expandedPaths:Array.from(this.state.expandedPaths)})),localStorage.setItem(this.projectKey,JSON.stringify({selectedPath:this.state.selectedPath,selectedPaths:Array.from(this.state.selectedPaths),targetPaths:Array.from(this.state.targetPaths),scrollTop:this.state.scrollTop,focusPathPerMode:this.state.focusPathPerMode}))}catch(e){console.warn("[TreeState] Failed to save state:",e)}}setupStorageListener(){window.addEventListener("storage",e=>{try{if(e.key===this.sharedKey&&e.newValue){const n=JSON.parse(e.newValue);this.state.expandedPaths=new Set(n.expandedPaths||[]),this.notifyListeners()}else if(e.key===this.projectKey&&e.newValue){const n=JSON.parse(e.newValue);this.state.selectedPath=n.selectedPath,this.state.selectedPaths=new Set(n.selectedPaths||[]),this.state.targetPaths=new Set(n.targetPaths||[]),this.state.scrollTop=n.scrollTop||0,this.state.focusPathPerMode=n.focusPathPerMode||to(),this.notifyListeners()}}catch(n){console.warn("[TreeState] Failed to parse storage event:",n)}})}notifyListeners(){this.listeners.forEach(e=>e(this.state))}subscribe(e){return this.listeners.add(e),()=>this.listeners.delete(e)}getState(){return this.state}isExpanded(e){return this.state.expandedPaths.has(e)}expand(e){this.state.expandedPaths.add(e),this.saveState(),this.notifyListeners()}collapse(e){this.state.expandedPaths.delete(e),this.saveState(),this.notifyListeners()}toggle(e){const n=this.isExpanded(e);return n?this.collapse(e):this.expand(e),!n}getExpanded(){return new Set(this.state.expandedPaths)}setSelected(e){this.state.selectedPath=e,this.saveState(),this.notifyListeners()}getSelected(){return this.state.selectedPath}isSelected(e){return this.state.selectedPaths.has(e)}addToSelection(e){this.state.selectedPaths.add(e),this.state.lastClickedPath=e,this.saveState(),this.notifyListeners()}removeFromSelection(e){this.state.selectedPaths.delete(e),this.saveState(),this.notifyListeners()}toggleSelection(e){this.state.selectedPaths.has(e)?this.state.selectedPaths.delete(e):this.state.selectedPaths.add(e),this.state.lastClickedPath=e,this.saveState(),this.notifyListeners()}getSelectedPaths(){return new Set(this.state.selectedPaths)}setSelectedPaths(e){this.state.selectedPaths=new Set(e),this.saveState(),this.notifyListeners()}clearSelection(){this.state.selectedPaths.clear(),this.state.selectedPath=null,this.state.lastClickedPath=null,this.saveState(),this.notifyListeners()}getLastClickedPath(){return this.state.lastClickedPath}setLastClickedPath(e){this.state.lastClickedPath=e}selectSingle(e){this.state.selectedPaths.clear(),this.state.selectedPaths.add(e),this.state.selectedPath=e,this.state.lastClickedPath=e,this.saveState(),this.notifyListeners()}selectSingleSilent(e){this.state.selectedPaths.clear(),this.state.selectedPaths.add(e),this.state.selectedPath=e,this.state.lastClickedPath=e,this.saveState()}setSelectedPathsSilent(e){this.state.selectedPaths=new Set(e),this.saveState()}setSelectedSilent(e){this.state.selectedPath=e,this.saveState()}setScrollTop(e){this.state.scrollTop=e,this.saveState()}getScrollTop(){return this.state.scrollTop}expandToPath(e){const n=e.split("/");let s="";for(let r=0;r<n.length-1;r++)s=s?`${s}/${n[r]}`:n[r],this.state.expandedPaths.add(s);this.saveState(),this.notifyListeners()}isTarget(e){return this.state.targetPaths.has(e)}addTarget(e){this.state.targetPaths.add(e),this.saveState(),this.notifyListeners()}removeTarget(e){this.state.targetPaths.delete(e),this.saveState(),this.notifyListeners()}setTargets(e){this.state.targetPaths=new Set(e),this.saveState(),this.notifyListeners()}clearTargets(){this.state.targetPaths.clear(),this.saveState(),this.notifyListeners()}getTargets(){return new Set(this.state.targetPaths)}setFocusPath(e,n){this.state.focusPathPerMode[e]=n,this.saveState(),this.notifyListeners()}getFocusPath(e){return this.state.focusPathPerMode[e]}clear(){this.state={expandedPaths:new Set,selectedPath:null,selectedPaths:new Set,targetPaths:new Set,scrollTop:0,focusPathPerMode:to(),lastClickedPath:null},localStorage.removeItem(this.projectKey),localStorage.removeItem(this.sharedKey),this.notifyListeners()}}const dr={code:{allowedExtensions:[],hiddenPatterns:["__pycache__",".pyc","node_modules",".git/objects"]},vis:{allowedExtensions:[".png",".jpg",".jpeg",".svg",".pdf",".csv",".json",".xlsx",".tsv"],hiddenPatterns:["__pycache__","node_modules",".git",".venv"]},writer:{allowedExtensions:[".tex",".bib",".pdf",".png",".jpg",".jpeg",".svg",".eps"],hiddenPatterns:["__pycache__","node_modules",".git",".venv","build",".aux",".log",".out"]},scholar:{allowedExtensions:[".bib"],hiddenPatterns:["__pycache__","node_modules",".git",".venv","build"]},clew:{allowedExtensions:[],hiddenPatterns:["__pycache__","node_modules",".git",".venv"]},hub:{allowedExtensions:[],hiddenPatterns:["__pycache__","node_modules",".git",".venv"]},files:{allowedExtensions:[],hiddenPatterns:["__pycache__","node_modules",".git"]},tools:{allowedExtensions:[],hiddenPatterns:["__pycache__","node_modules",".git",".venv"]},example:{allowedExtensions:[],hiddenPatterns:["__pycache__","node_modules",".git",".venv"]},apps:{allowedExtensions:[],hiddenPatterns:["__pycache__","node_modules",".git",".venv"]},explorer:{allowedExtensions:[],hiddenPatterns:["__pycache__","node_modules",".git",".venv"]},all:{allowedExtensions:[],disabledExtensions:[],hiddenPatterns:[]}},$v={code:["scripts"],vis:[],writer:[],scholar:[],clew:[],hub:[],files:[],tools:[],example:[],apps:[],explorer:[],all:[]},Rv={scholar:["scitex/scholar"],vis:["scitex/vis"],writer:["scitex/_writer"],code:[],clew:["scitex/verify"],hub:[],files:[],tools:[],explorer:[],example:[],apps:[],all:[]},ou={scholar:["node_modules",".git","__pycache__",".venv","venv","build","dist"],vis:["node_modules",".git","__pycache__",".venv","venv"],writer:["node_modules",".git","__pycache__",".venv","venv","build","dist","ai","config","docs","requirements","scripts","tests","texts","archive","figures","latex_styles","tables","wordcounts","logs","output"],code:[".git"],clew:["node_modules",".git","__pycache__",".venv","venv"],hub:["node_modules",".git","__pycache__",".venv","venv"],files:[".git"],tools:["node_modules",".git","__pycache__",".venv","venv"],explorer:["node_modules",".git","__pycache__",".venv","venv"],example:["node_modules",".git","__pycache__",".venv","venv"],apps:["node_modules",".git","__pycache__",".venv","venv"],all:[]},au={scholar:[".bib"],vis:[".csv",".tsv",".json",".xml",".png",".jpg",".jpeg",".gif",".svg",".webp",".pdf",".figz",".pltz"],writer:[".tex",".bib",".cls",".sty",".png",".jpg",".jpeg",".pdf",".svg",".eps",".csv",".tsv"],code:"all",clew:"all",hub:"all",files:"all",tools:"all",explorer:"all",example:"all",apps:"all",all:"all"},no=[".gitkeep"];class Av{constructor(e,n){this.showHidden=!1,this.moduleFilterEnabled=!1;const s=au[e],r=s==="all"?[]:s,i=ou[e]||[],o=dr[e]||dr.all;this.config={mode:e,allowedExtensions:(n==null?void 0:n.allowedExtensions)??r,disabledExtensions:(n==null?void 0:n.disabledExtensions)??o.disabledExtensions??[],hiddenPatterns:(n==null?void 0:n.hiddenPatterns)??i}}isHidden(e){const{name:n,type:s}=e;return!this.showHidden&&n.startsWith(".")?!0:s==="file"&&no.includes(n)?!1:!!(s==="file"&&[".DS_Store","Thumbs.db"].includes(n)||s==="directory"&&(n.endsWith(".figz.d")||n.endsWith(".pltz.d")))}isInactive(e){if(!this.moduleFilterEnabled)return!1;const{name:n,path:s,type:r}=e;for(const i of this.config.hiddenPatterns)if(i.startsWith(".")&&n.endsWith(i)||n===i||s.includes(`/${i}/`)||s.includes(`/${i}`))return!0;return!this.isWithinAllowedDirectories(s)||r==="file"&&!this.isAllowed(e)}isWithinAllowedDirectories(e){const n=Rv[this.config.mode];if(n.length===0)return!0;const s=e.replace(/^\.\//,"");return n.some(i=>{const o=i.replace(/^\.\//,"");return s.startsWith(o)||s===o})?!0:n.some(i=>i.replace(/^\.\//,"").startsWith(s+"/"))}isAllowed(e){if(e.type==="directory"||no.includes(e.name)||this.config.allowedExtensions.length===0)return!0;const n=this.getExtension(e.name);return this.config.allowedExtensions.includes(n)}isDisabled(e){if(!this.moduleFilterEnabled||e.type==="directory"||no.includes(e.name))return!1;const n=this.getExtension(e.name);return!!(this.config.disabledExtensions.includes(n)||this.config.allowedExtensions.length>0&&!this.config.allowedExtensions.includes(n))}getExtension(e){const n=e.lastIndexOf(".");return n===-1?"":e.substring(n).toLowerCase()}filterTree(e){return e.filter(n=>!this.isHidden(n)).map(n=>n.type==="directory"&&n.children?{...n,children:this.filterTree(n.children)}:n)}getMode(){return this.config.mode}getConfig(){return this.config}setAllowedExtensions(e){this.config.allowedExtensions=e}setDisabledExtensions(e){this.config.disabledExtensions=e}setHiddenPatterns(e){this.config.hiddenPatterns=e}setShowHidden(e){this.showHidden=e}getShowHidden(){return this.showHidden}setModuleFilterEnabled(e){this.moduleFilterEnabled=e}getModuleFilterEnabled(){return this.moduleFilterEnabled}setMode(e){const n=au[e],s=n==="all"?[]:n,r=ou[e]||[],i=dr[e]||dr.all;this.config={mode:e,allowedExtensions:s,disabledExtensions:i.disabledExtensions??[],hiddenPatterns:r}}}function jv(t){const e=t.substring(t.lastIndexOf(".")).toLowerCase(),n=t.toLowerCase(),s={".py":{icon:"fab fa-python",color:"#3776ab"},".js":{icon:"fab fa-js",color:"#f7df1e"},".ts":{icon:"fab fa-js",color:"#3178c6"},".jsx":{icon:"fab fa-react",color:"#61dafb"},".tsx":{icon:"fab fa-react",color:"#61dafb"},".java":{icon:"fab fa-java",color:"#007396"},".go":{icon:"fas fa-code",color:"#00add8"},".rs":{icon:"fas fa-gear",color:"#ce422b"},".c":{icon:"fas fa-file-code",color:"#555555"},".cpp":{icon:"fas fa-file-code",color:"#f34b7d"},".html":{icon:"fab fa-html5",color:"#e34c26"},".css":{icon:"fab fa-css3-alt",color:"#264de4"},".scss":{icon:"fab fa-sass",color:"#cc6699"},".json":{icon:"fas fa-brackets-curly",color:"#ffa500"},".yaml":{icon:"fas fa-file-code",color:"#cb171e"},".yml":{icon:"fas fa-file-code",color:"#cb171e"},".toml":{icon:"fas fa-file-code",color:"#9c4121"},".xml":{icon:"fas fa-file-code",color:"#e37933"},".csv":{icon:"fas fa-table",color:"#217346"},".md":{icon:"fab fa-markdown",color:"#000000"},".txt":{icon:"fas fa-file-alt",color:"#777777"},".pdf":{icon:"fas fa-file-pdf",color:"#ff0000"},".tex":{icon:"fas fa-file-alt",color:"#3d6117"},".bib":{icon:"fas fa-book-open",color:"#cb4b16"},".sh":{icon:"fas fa-terminal",color:"#89e051"},".bash":{icon:"fas fa-terminal",color:"#89e051"},".r":{icon:"fab fa-r-project",color:"#276dc3"},".ipynb":{icon:"fas fa-book",color:"#ff6f00"},".sql":{icon:"fas fa-database",color:"#e38c00"},".jpg":{icon:"fas fa-image",color:"#00bfa5"},".jpeg":{icon:"fas fa-image",color:"#00bfa5"},".png":{icon:"fas fa-image",color:"#00bfa5"},".gif":{icon:"fas fa-image",color:"#00bfa5"},".svg":{icon:"fas fa-image",color:"#ffb300"},".webp":{icon:"fas fa-image",color:"#00bfa5"},".zip":{icon:"fas fa-file-archive",color:"#8b8b8b"},".tar":{icon:"fas fa-file-archive",color:"#8b8b8b"},".gz":{icon:"fas fa-file-archive",color:"#8b8b8b"},".log":{icon:"fas fa-file-lines",color:"#777777"},".env":{icon:"fas fa-cog",color:"#edb92e"}},r={"readme.md":{icon:"fas fa-book",color:"#00b0ff"},readme:{icon:"fas fa-book",color:"#00b0ff"},license:{icon:"fas fa-scroll",color:"#ffab00"},dockerfile:{icon:"fab fa-docker",color:"#2496ed"},makefile:{icon:"fas fa-cog",color:"#6d6d6d"},".gitignore":{icon:"fab fa-git-alt",color:"#f05032"},".gitattributes":{icon:"fab fa-git-alt",color:"#f05032"},"package.json":{icon:"fab fa-npm",color:"#cb3837"},"tsconfig.json":{icon:"fab fa-js",color:"#3178c6"},"requirements.txt":{icon:"fab fa-python",color:"#3776ab"},"setup.py":{icon:"fab fa-python",color:"#3776ab"}};if(r[n]){const{icon:i,color:o}=r[n];return`<i class="${i}" style="color: ${o}; width: 16px; text-align: center;"></i>`}if(s[e]){const{icon:i,color:o}=s[e];return`<i class="${i}" style="color: ${o}; width: 16px; text-align: center;"></i>`}return'<i class="fas fa-file" style="color: #777; width: 16px; text-align: center;"></i>'}function zv(){return'<i class="fas fa-folder" style="color: #dcb67a; width: 16px; text-align: center;"></i>'}class Hv{constructor(e,n,s){this.searchMatches=new Set,this.searchAncestors=new Set,this.searchActive=!1,this.sortMode="name",this.config=e,this.stateManager=n,this.filter=s}renderLoadingSkeleton(){return`
       <div class="wft-tree wft-loading">
         <div class="wft-skeleton-root">
           <span class="wft-skeleton-icon folder"></span>
@@ -384,7 +384,7 @@ __pycache__/*
     <span>I can take actions: stats, plots, literature, and your current work.</span>
   </div>`;class Ox{constructor(){this.messagesEl=null,this.inputEl=null,this.sendBtn=null,this.speakBtn=null,this.micBtn=null,this.sttModelSelect=null,this.modelBadge=null,this.busy=!1,this.autoSpeak=!1,this.currentAudio=null,this.recorder=null,this.imageInput=null,this.sketchCanvas=null,this.webcamCapture=null,this.history=[],this.historyIdx=-1,this.historyDraft="",this.context={},this.sessionsPanel=null,this._userAtBottom=!0,this.emptyStateHtml=Hx}init(e,n,s){var r,i,o,a;if(this.messagesEl=e.messagesEl,this.inputEl=e.inputEl,this.sendBtn=e.sendBtn,this.speakBtn=e.speakBtn,this.micBtn=e.micBtn,this.sttModelSelect=e.sttModelSelect,this.modelBadge=e.modelBadge,this.adapter=n.adapter,this.sttAdapter=n.sttAdapter,this.ttsAdapter=n.ttsAdapter,this.autoSpeak=n.autoSpeak??!1,n.emptyStateHtml!==void 0&&(this.emptyStateHtml=n.emptyStateHtml),s&&(this.context=s),this.recorder=new Fx(e.volBars,this.micBtn),(r=this.micBtn)==null||r.addEventListener("click",()=>this.toggleRecording()),this.history=Px(),e.imagePreviewEl&&e.imageFileInput&&(this.imageInput=new Dx(e.imagePreviewEl,e.imageFileInput),this.inputEl&&this.imageInput.bindPaste(this.inputEl),this.sketchCanvas=new jx(this.imageInput),this.webcamCapture=new $x(this.imageInput,e.imageFileInput),(i=e.cameraBtn)==null||i.addEventListener("click",()=>{var l;return void((l=this.webcamCapture)==null?void 0:l.open())}),(o=e.sketchBtn)==null||o.addEventListener("click",()=>{var l;return(l=this.sketchCanvas)==null?void 0:l.open()})),(a=this.messagesEl)==null||a.addEventListener("scroll",()=>{if(!this.messagesEl)return;const l=this.messagesEl;this._userAtBottom=l.scrollHeight-l.scrollTop-l.clientHeight<40}),this.messagesEl){const l=this.messagesEl;new MutationObserver(()=>{this._userAtBottom&&(l.scrollTop=999999)}).observe(l,{childList:!0,subtree:!0})}}scrollToBottomIfNeeded(){this.messagesEl&&(this.messagesEl.scrollTop=999999)}scrollToBottom(){this.messagesEl&&(this._userAtBottom=!0,this.messagesEl.scrollTop=999999)}setContext(e){this.context=e}getContext(){return this.context}setSessionsPanel(e){this.sessionsPanel=e}clearChat(){this.clearConversation()}clearConversation(){xx(),this.messagesEl&&(this.messagesEl.innerHTML=this.emptyStateHtml)}restoreConversation(){var n;const e=af();if(!(e.length===0||!this.messagesEl)){this.messagesEl.innerHTML="";for(const s of e){const r=this.createMsgEl(s.role);if(s.role==="assistant"&&s.text){const i=document.createElement("div");i.className="stx-shell-ai-md-segment",i.innerHTML=Zr(s.text),la(i),ca(i),r.appendChild(i)}else r.appendChild(document.createTextNode(s.text));(n=s.toolsUsed)!=null&&n.length&&aa(r,s.toolsUsed)}this.scrollToBottom(),this._scrollRepeatedlyForLayout()}}loadSessionMessages(e){var n;if(this.messagesEl){this.messagesEl.innerHTML="";for(const s of e){const r=this.createMsgEl(s.role);if(s.role==="assistant"&&s.text){const i=document.createElement("div");i.className="stx-shell-ai-md-segment",i.innerHTML=Zr(s.text),la(i),ca(i),r.appendChild(i)}else r.appendChild(document.createTextNode(s.text));(n=s.tools_used)!=null&&n.length&&aa(r,s.tools_used)}this.scrollToBottom(),this._scrollRepeatedlyForLayout()}}async copyChat(){if(!this.messagesEl)return;const e=this.messagesEl.querySelectorAll(".stx-shell-ai-msg"),n=[];e.forEach(r=>{const i=r.classList.contains("user")?"You":"AI";n.push(`${i}: ${r.innerText.trim()}`)});const s=n.join(`
 
-`);try{await navigator.clipboard.writeText(s)}catch{}}printChat(){this.messagesEl&&(document.body.classList.add("scitex-print-chat"),window.print(),document.body.classList.remove("scitex-print-chat"))}createMsgEl(e){var s;const n=document.createElement("div");return n.className=`stx-shell-ai-msg ${e}`,(s=this.messagesEl)==null||s.appendChild(n),this.scrollToBottomIfNeeded(),n}navigateHistory(e){if(!this.inputEl||this.history.length===0)return;this.historyIdx===-1&&(this.historyDraft=this.inputEl.value);const n=this.historyIdx+e;if(n<-1||n>=this.history.length)return;this.historyIdx=n,this.inputEl.value=n===-1?this.historyDraft:this.history[n],this.inputEl.dispatchEvent(new Event("input"));const s=this.inputEl.value.length;this.inputEl.setSelectionRange(s,s)}collectPageHints(){const e=[];document.querySelectorAll("[data-ai-hint]").forEach(r=>{const i=r.dataset.aiHint;i&&e.push(i)});const n=document.getElementById("ws-viewer-sidebar"),s=n==null?void 0:n.dataset.aiViewerActive;return s&&e.push(`Currently open in editor: ${s}`),e}async send(){var r,i,o,a,l,c;if(this.busy||!this.inputEl||!this.messagesEl)return;const e=this.inputEl.value.trim();if(!e)return;if(this.history=Lx(this.history,e),this.historyIdx=-1,e==="/clear"){this.clearChat(),this.inputEl.value="";return}(r=this.currentAudio)==null||r.pause(),this.currentAudio=null,(i=this.messagesEl.querySelector(".stx-shell-ai-empty"))==null||i.remove(),this._userAtBottom=!0;const n=this.createMsgEl("user");n.textContent=e,(o=this.imageInput)==null||o.renderInlineThumbsInto(n),Hs({role:"user",text:e}),(a=this.sessionsPanel)==null||a.saveMessage("user",e);const s=(l=this.imageInput)!=null&&l.hasAttachments()?this.imageInput.getAttachmentsAsBase64():[];(c=this.imageInput)==null||c.clearAttachments(),this.inputEl.value="",this.inputEl.style.height="auto",this.busy=!0,this.sendBtn&&(this.sendBtn.disabled=!0),this.context.page_hints=this.collectPageHints();try{await zx({prompt:e,adapter:this.adapter,context:this.context,images:s,messagesEl:this.messagesEl,modelBadge:this.modelBadge,autoSpeak:this.autoSpeak,speak:d=>void this.speak(d),scrollIfNeeded:()=>this.scrollToBottomIfNeeded(),createMsgEl:d=>this.createMsgEl(d),onAssistantDone:(d,u)=>{var f;this.scrollToBottom(),setTimeout(()=>this.scrollToBottom(),200),setTimeout(()=>this.scrollToBottom(),500),u&&((f=this.sessionsPanel)==null||f.saveMessage("assistant",u))}})}finally{this.busy=!1,this.sendBtn&&(this.sendBtn.disabled=!1)}}async speak(e){var n;(n=this.currentAudio)==null||n.pause(),this.currentAudio=null,this.currentAudio=await _x(e,this.ttsAdapter)}toggleSpeak(){var e,n;this.autoSpeak=!this.autoSpeak,(e=this.speakBtn)==null||e.classList.toggle("active",this.autoSpeak),this.autoSpeak||((n=this.currentAudio)==null||n.pause(),this.currentAudio=null)}get isAutoSpeak(){return this.autoSpeak}toggleRecording(){!this.recorder||!this.sttAdapter||(this.recorder.isRecording?this.recorder.stop():this.recorder.start(this.sttAdapter,e=>{if(!this.inputEl)return;const n=this.inputEl.value.trim();this.inputEl.value=n?`${n} ${e}`:e,this.inputEl.dispatchEvent(new Event("input")),this.inputEl.focus()},()=>{var e;return((e=this.sttModelSelect)==null?void 0:e.value)??""}))}get isBusy(){return this.busy}_scrollRepeatedlyForLayout(){const e=setInterval(()=>{this.messagesEl&&(this.messagesEl.scrollTop=999999)},100);setTimeout(()=>clearInterval(e),3e3)}}class Bx{constructor(){this.currentSessionId=null,this.sessions=[],this.listEl=null,this.chatCounter=0,this.contextMenu=null,this.adapter=null,this.onSwitch=null,this.onClear=null}init(e,n,s,r){this.listEl=e,this.adapter=n,this.onSwitch=s,this.onClear=r;const i=document.querySelector(".stx-shell-ai-new-chat");i==null||i.addEventListener("click",()=>this.newChat());const o=sessionStorage.getItem("scitex_ai_session_id");o&&(this.currentSessionId=parseInt(o,10)),document.addEventListener("click",()=>this.dismissContextMenu()),this.loadSessions()}async loadSessions(){if(this.adapter)try{const e=await this.adapter.listSessions();this.sessions=e,!this.currentSessionId&&e.length>0&&(this.currentSessionId=e[0].id,sessionStorage.setItem("scitex_ai_session_id",String(e[0].id))),this.render(e),e.length===0&&await this.createSession("C1")}catch(e){console.error("[SessionsPanel] loadSessions failed:",e)}}async createSession(e){if(!this.adapter)return null;try{const n=await this.adapter.createSession(e||`C${this.chatCounter+1}`);return this.currentSessionId=n.id,sessionStorage.setItem("scitex_ai_session_id",String(n.id)),this.loadSessions(),n.id}catch(n){return console.error("[SessionsPanel] createSession failed:",n),null}}async switchSession(e){var n;if(this.adapter)try{const s=await this.adapter.getMessages(e);this.currentSessionId=e,sessionStorage.setItem("scitex_ai_session_id",String(e)),(n=this.onSwitch)==null||n.call(this,s.messages,e),this.highlightActive()}catch(s){console.error("[SessionsPanel] switchSession failed:",s)}}async saveMessage(e,n){if(!(!this.currentSessionId||!this.adapter))try{await this.adapter.addMessage(this.currentSessionId,e,n),this.loadSessions()}catch(s){console.error("[SessionsPanel] saveMessage failed:",s)}}newChat(){var e;this.currentSessionId=null,sessionStorage.removeItem("scitex_ai_session_id"),(e=this.onClear)==null||e.call(this),this.highlightActive()}getSessionId(){return this.currentSessionId}render(e){if(this.listEl){this.listEl.innerHTML="",this.chatCounter=e.length;for(const n of e){const s=document.createElement("div");s.className="stx-shell-ai-session-item",n.id===this.currentSessionId&&s.classList.add("active"),s.dataset.sessionId=String(n.id);const r=document.createElement("span");r.className="stx-shell-ai-session-title",r.textContent=n.title,r.addEventListener("dblclick",o=>{o.stopPropagation(),this.startRename(s,n.id,n.title)});const i=document.createElement("button");i.className="stx-shell-ai-session-del",i.innerHTML='<i class="fas fa-times"></i>',i.title="Delete",i.addEventListener("click",o=>{o.stopPropagation(),this.deleteSession(n.id)}),s.appendChild(r),s.appendChild(i),s.addEventListener("click",()=>void this.switchSession(n.id)),s.addEventListener("contextmenu",o=>{o.preventDefault(),o.stopPropagation(),this.showContextMenu(o,n)}),this.listEl.appendChild(s)}}}showContextMenu(e,n){this.dismissContextMenu();const s=document.createElement("div");s.className="stx-shell-ai-context-menu",s.style.left=`${e.clientX}px`,s.style.top=`${e.clientY}px`;const r=[{label:"Rename",icon:"fas fa-pen",action:()=>{var o;const i=(o=this.listEl)==null?void 0:o.querySelector(`[data-session-id="${n.id}"]`);i&&this.startRename(i,n.id,n.title)}},{label:"Delete",icon:"fas fa-trash",action:()=>void this.deleteSession(n.id)}];for(const i of r){const o=document.createElement("div");o.className="stx-shell-ai-context-menu-item",o.innerHTML=`<i class="${i.icon}"></i> ${i.label}`,o.addEventListener("click",a=>{a.stopPropagation(),this.dismissContextMenu(),i.action()}),s.appendChild(o)}document.body.appendChild(s),this.contextMenu=s}dismissContextMenu(){this.contextMenu&&(this.contextMenu.remove(),this.contextMenu=null)}highlightActive(){if(this.listEl)for(const e of this.listEl.querySelectorAll(".stx-shell-ai-session-item")){const n=parseInt(e.dataset.sessionId||"0",10);e.classList.toggle("active",n===this.currentSessionId)}}startRename(e,n,s){const r=e.querySelector(".stx-shell-ai-session-title");if(!r)return;const i=document.createElement("input");i.className="stx-shell-ai-session-rename",i.value=s,r.replaceWith(i),i.focus(),i.select();const o=async()=>{var l;const a=i.value.trim()||s;if(i.replaceWith(r),r.textContent=a,a!==s&&((l=this.adapter)!=null&&l.renameSession))try{await this.adapter.renameSession(n,a)}catch(c){console.error("[SessionsPanel] rename failed:",c)}};i.addEventListener("blur",()=>void o()),i.addEventListener("keydown",a=>{a.key==="Enter"?(a.preventDefault(),i.blur()):a.key==="Escape"&&(i.value=s,i.blur())})}async deleteSession(e){if(this.adapter)try{await this.adapter.deleteSession(e),this.currentSessionId===e&&this.newChat(),this.loadSessions()}catch(n){console.error("[SessionsPanel] deleteSession failed:",n)}}}const Su=document.getElementById("root"),Ux=new URLSearchParams(window.location.search),Wx=Ux.get("mode")==="embedded";Su&&(document.documentElement.setAttribute("data-theme","dark"),ao.createRoot(Su).render(h.jsx(fs.StrictMode,{children:h.jsx(By,{embedded:Wx})})));const Kx={async fetchTree(){const t=await fetch("api/tree");return t.ok?{success:!0,tree:(await t.json()).tree??[]}:{success:!1,tree:[],error:`HTTP ${t.status}: ${t.statusText}`}},getCsrfToken(){const t=document.querySelector('meta[name="csrf-token"]');return(t==null?void 0:t.content)??""}},cf=new xw({mode:"files",containerId:"ws-worktree-tree",ownerUsername:"figrecipe",projectSlug:"figrecipe",adapter:Kx,showGitStatus:!1,showFolderActions:!1,onFileSelect:(t,e)=>{window.dispatchEvent(new CustomEvent("figrecipe:file-select",{detail:{path:e.path,name:e.name}}))}});cf.initialize();const Vx={async fetchRecentFiles(){const t=await fetch("api/files?recent=true&limit=50");if(!t.ok)return[];const e=await t.json();return(e.files??e??[]).map(s=>({path:s.path,name:s.name??s.path.split("/").pop()??s.path,modified_at:s.modified_at??s.mtime??new Date().toISOString(),action:s.action}))},getFileUrl(t){return`api/file-content/${t}?raw=true`}};Ow();Bw({adapter:Vx,pollIntervalMs:1e4});const Xx={getWebSocketUrl(){return`ws://127.0.0.1:${parseInt(window.location.port||"5050",10)+1}/`}};Iv({container:"#stx-shell-ai-console-terminal",adapter:Xx,clipboard:!0});const Yx=new Ew;Yx.init();const Gx=new Cw;Gx.init();Lw();sf("figrecipe",[{title:"Figure Editor",shortcuts:[{keys:"Ctrl+Z",description:"Undo"},{keys:"Ctrl+Y",description:"Redo"},{keys:"Del",description:"Delete selected"}]}]);const qx={async readFile(t){const e=await fetch(`api/file-content/${t}`);if(!e.ok)throw new Error(`HTTP ${e.status}`);return{content:(await e.json()).content??""}},getFileUrl(t){return`api/file-content/${t}?raw=true`}};let ua=null;try{ua=new yx({adapter:qx})}catch{}window.addEventListener("figrecipe:file-select",t=>{var e;ua&&((e=t.detail)!=null&&e.path)&&ua.openFile(t.detail.path)});const Qx="You are a helpful AI assistant in the FigRecipe figure editor. Help with YAML recipes, matplotlib plots, and figure composition.",Zx={async streamChat(t,e,n){return fetch("api/chat/stream",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({prompt:t,history:[],system_prompt:Qx,...n!=null&&n.length?{images:n.map(s=>`data:image/png;base64,${s}`)}:{}})})}},Jx=document.getElementById("stx-shell-ai-image-preview"),e0=document.getElementById("stx-shell-ai-file-input"),gt=new Ox;gt.init({messagesEl:document.getElementById("stx-shell-ai-messages"),inputEl:document.getElementById("stx-shell-ai-input"),sendBtn:null,speakBtn:null,micBtn:document.getElementById("stx-shell-ai-mic"),sttModelSelect:null,modelBadge:document.getElementById("stx-shell-ai-model-badge"),volBars:Array.from(document.querySelectorAll(".stx-shell-ai-vol-bar")),imagePreviewEl:Jx,imageFileInput:e0,cameraBtn:document.getElementById("stx-shell-ai-camera"),sketchBtn:document.getElementById("stx-shell-ai-sketch")},{adapter:Zx,sttAdapter:{async transcribe(t){return console.log("[figrecipe] Voice recorded, size:",t.size),""}}});gt.restoreConversation();const t0={async listSessions(){const t=await fetch("api/chat/sessions/");if(!t.ok)throw new Error(`HTTP ${t.status}`);return(await t.json()).sessions??[]},async getMessages(t){const e=await fetch(`api/chat/sessions/${t}/messages/`);if(!e.ok)throw new Error(`HTTP ${e.status}`);return e.json()},async createSession(t){const e=await fetch("api/chat/sessions/",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({title:t||"New Chat"})});if(!e.ok)throw new Error(`HTTP ${e.status}`);return e.json()},async deleteSession(t){const e=await fetch(`api/chat/sessions/${t}/`,{method:"DELETE"});if(!e.ok)throw new Error(`HTTP ${e.status}`)},async addMessage(t,e,n){const s=await fetch(`api/chat/sessions/${t}/messages/`,{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({role:e,content:n})});if(!s.ok)throw new Error(`HTTP ${s.status}`)},async renameSession(t,e){const n=await fetch(`api/chat/sessions/${t}/`,{method:"PATCH",headers:{"Content-Type":"application/json"},body:JSON.stringify({title:e})});if(!n.ok)throw new Error(`HTTP ${n.status}`)}},Eu=new Bx,ku=document.getElementById("stx-shell-ai-sessions-list");ku&&(Eu.init(ku,t0,(t,e)=>{gt.loadSessionMessages(t.map(n=>({role:n.role,text:n.content,tools_used:[],media:[]})))},()=>{gt.clearChat()}),gt.setSessionsPanel(Eu));const yt=document.getElementById("stx-shell-ai-input");yt&&(yt.addEventListener("keydown",t=>{if(t.key==="Enter"&&!t.shiftKey){t.preventDefault(),gt.send();return}if((t.key==="ArrowUp"||t.ctrlKey&&t.key==="p")&&yt.selectionStart===0){t.preventDefault(),gt.navigateHistory(1);return}if((t.key==="ArrowDown"||t.ctrlKey&&t.key==="n")&&yt.selectionEnd===yt.value.length){t.preventDefault(),gt.navigateHistory(-1);return}}),yt.addEventListener("input",()=>{yt.style.height="auto",yt.style.height=Math.min(yt.scrollHeight,150)+"px"}));document.addEventListener("stx-shell:clear-chat",()=>{gt.clearChat()});window.addEventListener("stx-shell:camera",()=>{const t=document.querySelector('.stx-shell-ai-input-btn[title="Camera"]');t==null||t.click()});window.addEventListener("stx-shell:sketch",()=>{const t=document.querySelector('.stx-shell-ai-input-btn[title="Sketch"]');t==null||t.click()});window.addEventListener("stx-shell:mic-toggle",()=>{gt.toggleRecording()});let bu=!1;window.addEventListener("stx-shell:settings",()=>{if(bu)return;bu=!0;const t=document.getElementById("ai-agent-sources-content");t&&(t.innerHTML=`
+`);try{await navigator.clipboard.writeText(s)}catch{}}printChat(){this.messagesEl&&(document.body.classList.add("scitex-print-chat"),window.print(),document.body.classList.remove("scitex-print-chat"))}createMsgEl(e){var s;const n=document.createElement("div");return n.className=`stx-shell-ai-msg ${e}`,(s=this.messagesEl)==null||s.appendChild(n),this.scrollToBottomIfNeeded(),n}navigateHistory(e){if(!this.inputEl||this.history.length===0)return;this.historyIdx===-1&&(this.historyDraft=this.inputEl.value);const n=this.historyIdx+e;if(n<-1||n>=this.history.length)return;this.historyIdx=n,this.inputEl.value=n===-1?this.historyDraft:this.history[n],this.inputEl.dispatchEvent(new Event("input"));const s=this.inputEl.value.length;this.inputEl.setSelectionRange(s,s)}collectPageHints(){const e=[];document.querySelectorAll("[data-ai-hint]").forEach(r=>{const i=r.dataset.aiHint;i&&e.push(i)});const n=document.getElementById("ws-viewer-sidebar"),s=n==null?void 0:n.dataset.aiViewerActive;return s&&e.push(`Currently open in editor: ${s}`),e}async send(){var r,i,o,a,l,c;if(this.busy||!this.inputEl||!this.messagesEl)return;const e=this.inputEl.value.trim();if(!e)return;if(this.history=Lx(this.history,e),this.historyIdx=-1,e==="/clear"){this.clearChat(),this.inputEl.value="";return}(r=this.currentAudio)==null||r.pause(),this.currentAudio=null,(i=this.messagesEl.querySelector(".stx-shell-ai-empty"))==null||i.remove(),this._userAtBottom=!0;const n=this.createMsgEl("user");n.textContent=e,(o=this.imageInput)==null||o.renderInlineThumbsInto(n),Hs({role:"user",text:e}),(a=this.sessionsPanel)==null||a.saveMessage("user",e);const s=(l=this.imageInput)!=null&&l.hasAttachments()?this.imageInput.getAttachmentsAsBase64():[];(c=this.imageInput)==null||c.clearAttachments(),this.inputEl.value="",this.inputEl.style.height="auto",this.busy=!0,this.sendBtn&&(this.sendBtn.disabled=!0),this.context.page_hints=this.collectPageHints();try{await zx({prompt:e,adapter:this.adapter,context:this.context,images:s,messagesEl:this.messagesEl,modelBadge:this.modelBadge,autoSpeak:this.autoSpeak,speak:d=>void this.speak(d),scrollIfNeeded:()=>this.scrollToBottomIfNeeded(),createMsgEl:d=>this.createMsgEl(d),onAssistantDone:(d,u)=>{var f;this.scrollToBottom(),setTimeout(()=>this.scrollToBottom(),200),setTimeout(()=>this.scrollToBottom(),500),u&&((f=this.sessionsPanel)==null||f.saveMessage("assistant",u))}})}finally{this.busy=!1,this.sendBtn&&(this.sendBtn.disabled=!1)}}async speak(e){var n;(n=this.currentAudio)==null||n.pause(),this.currentAudio=null,this.currentAudio=await _x(e,this.ttsAdapter)}toggleSpeak(){var e,n;this.autoSpeak=!this.autoSpeak,(e=this.speakBtn)==null||e.classList.toggle("active",this.autoSpeak),this.autoSpeak||((n=this.currentAudio)==null||n.pause(),this.currentAudio=null)}get isAutoSpeak(){return this.autoSpeak}toggleRecording(){!this.recorder||!this.sttAdapter||(this.recorder.isRecording?this.recorder.stop():this.recorder.start(this.sttAdapter,e=>{if(!this.inputEl)return;const n=this.inputEl.value.trim();this.inputEl.value=n?`${n} ${e}`:e,this.inputEl.dispatchEvent(new Event("input")),this.inputEl.focus()},()=>{var e;return((e=this.sttModelSelect)==null?void 0:e.value)??""}))}get isBusy(){return this.busy}_scrollRepeatedlyForLayout(){const e=setInterval(()=>{this.messagesEl&&(this.messagesEl.scrollTop=999999)},100);setTimeout(()=>clearInterval(e),3e3)}}class Bx{constructor(){this.currentSessionId=null,this.sessions=[],this.listEl=null,this.chatCounter=0,this.contextMenu=null,this.adapter=null,this.onSwitch=null,this.onClear=null}init(e,n,s,r){this.listEl=e,this.adapter=n,this.onSwitch=s,this.onClear=r;const i=document.querySelector(".stx-shell-ai-new-chat");i==null||i.addEventListener("click",()=>this.newChat());const o=sessionStorage.getItem("scitex_ai_session_id");o&&(this.currentSessionId=parseInt(o,10)),document.addEventListener("click",()=>this.dismissContextMenu()),this.loadSessions()}async loadSessions(){if(this.adapter)try{const e=await this.adapter.listSessions();this.sessions=e,!this.currentSessionId&&e.length>0&&(this.currentSessionId=e[0].id,sessionStorage.setItem("scitex_ai_session_id",String(e[0].id))),this.render(e),e.length===0&&await this.createSession("C1")}catch(e){console.error("[SessionsPanel] loadSessions failed:",e)}}async createSession(e){if(!this.adapter)return null;try{const n=await this.adapter.createSession(e||`C${this.chatCounter+1}`);return this.currentSessionId=n.id,sessionStorage.setItem("scitex_ai_session_id",String(n.id)),this.loadSessions(),n.id}catch(n){return console.error("[SessionsPanel] createSession failed:",n),null}}async switchSession(e){var n;if(this.adapter)try{const s=await this.adapter.getMessages(e);this.currentSessionId=e,sessionStorage.setItem("scitex_ai_session_id",String(e)),(n=this.onSwitch)==null||n.call(this,s.messages,e),this.highlightActive()}catch(s){console.error("[SessionsPanel] switchSession failed:",s)}}async saveMessage(e,n){if(!(!this.currentSessionId||!this.adapter))try{await this.adapter.addMessage(this.currentSessionId,e,n),this.loadSessions()}catch(s){console.error("[SessionsPanel] saveMessage failed:",s)}}newChat(){var e;this.currentSessionId=null,sessionStorage.removeItem("scitex_ai_session_id"),(e=this.onClear)==null||e.call(this),this.highlightActive()}getSessionId(){return this.currentSessionId}render(e){if(this.listEl){this.listEl.innerHTML="",this.chatCounter=e.length;for(const n of e){const s=document.createElement("div");s.className="stx-shell-ai-session-item",n.id===this.currentSessionId&&s.classList.add("active"),s.dataset.sessionId=String(n.id);const r=document.createElement("span");r.className="stx-shell-ai-session-title",r.textContent=n.title,r.addEventListener("dblclick",o=>{o.stopPropagation(),this.startRename(s,n.id,n.title)});const i=document.createElement("button");i.className="stx-shell-ai-session-del",i.innerHTML='<i class="fas fa-times"></i>',i.title="Delete",i.addEventListener("click",o=>{o.stopPropagation(),this.deleteSession(n.id)}),s.appendChild(r),s.appendChild(i),s.addEventListener("click",()=>void this.switchSession(n.id)),s.addEventListener("contextmenu",o=>{o.preventDefault(),o.stopPropagation(),this.showContextMenu(o,n)}),this.listEl.appendChild(s)}}}showContextMenu(e,n){this.dismissContextMenu();const s=document.createElement("div");s.className="stx-shell-ai-context-menu",s.style.left=`${e.clientX}px`,s.style.top=`${e.clientY}px`;const r=[{label:"Rename",icon:"fas fa-pen",action:()=>{var o;const i=(o=this.listEl)==null?void 0:o.querySelector(`[data-session-id="${n.id}"]`);i&&this.startRename(i,n.id,n.title)}},{label:"Delete",icon:"fas fa-trash",action:()=>void this.deleteSession(n.id)}];for(const i of r){const o=document.createElement("div");o.className="stx-shell-ai-context-menu-item",o.innerHTML=`<i class="${i.icon}"></i> ${i.label}`,o.addEventListener("click",a=>{a.stopPropagation(),this.dismissContextMenu(),i.action()}),s.appendChild(o)}document.body.appendChild(s),this.contextMenu=s}dismissContextMenu(){this.contextMenu&&(this.contextMenu.remove(),this.contextMenu=null)}highlightActive(){if(this.listEl)for(const e of this.listEl.querySelectorAll(".stx-shell-ai-session-item")){const n=parseInt(e.dataset.sessionId||"0",10);e.classList.toggle("active",n===this.currentSessionId)}}startRename(e,n,s){const r=e.querySelector(".stx-shell-ai-session-title");if(!r)return;const i=document.createElement("input");i.className="stx-shell-ai-session-rename",i.value=s,r.replaceWith(i),i.focus(),i.select();const o=async()=>{var l;const a=i.value.trim()||s;if(i.replaceWith(r),r.textContent=a,a!==s&&((l=this.adapter)!=null&&l.renameSession))try{await this.adapter.renameSession(n,a)}catch(c){console.error("[SessionsPanel] rename failed:",c)}};i.addEventListener("blur",()=>void o()),i.addEventListener("keydown",a=>{a.key==="Enter"?(a.preventDefault(),i.blur()):a.key==="Escape"&&(i.value=s,i.blur())})}async deleteSession(e){if(this.adapter)try{await this.adapter.deleteSession(e),this.currentSessionId===e&&this.newChat(),this.loadSessions()}catch(n){console.error("[SessionsPanel] deleteSession failed:",n)}}}const Ux="You are a helpful AI assistant in the FigRecipe figure editor. Help with YAML recipes, matplotlib plots, and figure composition.",Wx={async streamChat(t,e,n){return fetch("api/chat/stream",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({prompt:t,history:[],system_prompt:Ux,...n!=null&&n.length?{images:n.map(s=>`data:image/png;base64,${s}`)}:{}})})}},Su=document.getElementById("root"),Kx=new URLSearchParams(window.location.search),Vx=Kx.get("mode")==="embedded";Su&&(document.documentElement.setAttribute("data-theme","dark"),ao.createRoot(Su).render(h.jsx(fs.StrictMode,{children:h.jsx(By,{embedded:Vx})})));const Xx={async fetchTree(){const t=await fetch("api/tree");return t.ok?{success:!0,tree:(await t.json()).tree??[]}:{success:!1,tree:[],error:`HTTP ${t.status}: ${t.statusText}`}},getCsrfToken(){const t=document.querySelector('meta[name="csrf-token"]');return(t==null?void 0:t.content)??""}},cf=new xw({mode:"files",containerId:"ws-worktree-tree",ownerUsername:"figrecipe",projectSlug:"figrecipe",adapter:Xx,showGitStatus:!1,showFolderActions:!1,onFileSelect:(t,e)=>{window.dispatchEvent(new CustomEvent("figrecipe:file-select",{detail:{path:e.path,name:e.name}}))}});cf.initialize();const Yx={async fetchRecentFiles(){const t=await fetch("api/files?recent=true&limit=50");if(!t.ok)return[];const e=await t.json();return(e.files??e??[]).map(s=>({path:s.path,name:s.name??s.path.split("/").pop()??s.path,modified_at:s.modified_at??s.mtime??new Date().toISOString(),action:s.action}))},getFileUrl(t){return`api/file-content/${t}?raw=true`}};Ow();Bw({adapter:Yx,pollIntervalMs:1e4});const Gx={getWebSocketUrl(){return`ws://127.0.0.1:${parseInt(window.location.port||"5050",10)+1}/`}};Iv({container:"#stx-shell-ai-console-terminal",adapter:Gx,clipboard:!0});const qx=new Ew;qx.init();const Qx=new Cw;Qx.init();Lw();sf("figrecipe",[{title:"Figure Editor",shortcuts:[{keys:"Ctrl+Z",description:"Undo"},{keys:"Ctrl+Y",description:"Redo"},{keys:"Del",description:"Delete selected"}]}]);const Zx={async readFile(t){const e=await fetch(`api/file-content/${t}`);if(!e.ok)throw new Error(`HTTP ${e.status}`);return{content:(await e.json()).content??""}},getFileUrl(t){return`api/file-content/${t}?raw=true`}};let ua=null;try{ua=new yx({adapter:Zx})}catch{}window.addEventListener("figrecipe:file-select",t=>{var e;ua&&((e=t.detail)!=null&&e.path)&&ua.openFile(t.detail.path)});window.addEventListener("figrecipe:file-select",t=>{var n;const e=(n=t.detail)==null?void 0:n.path;e&&/\.ya?ml$/i.test(e)&&$.getState().switchFile(e)});const Jx=document.getElementById("stx-shell-ai-image-preview"),e0=document.getElementById("stx-shell-ai-file-input"),gt=new Ox;gt.init({messagesEl:document.getElementById("stx-shell-ai-messages"),inputEl:document.getElementById("stx-shell-ai-input"),sendBtn:null,speakBtn:null,micBtn:document.getElementById("stx-shell-ai-mic"),sttModelSelect:null,modelBadge:document.getElementById("stx-shell-ai-model-badge"),volBars:Array.from(document.querySelectorAll(".stx-shell-ai-vol-bar")),imagePreviewEl:Jx,imageFileInput:e0,cameraBtn:document.getElementById("stx-shell-ai-camera"),sketchBtn:document.getElementById("stx-shell-ai-sketch")},{adapter:Wx,sttAdapter:{async transcribe(t){return console.log("[figrecipe] Voice recorded, size:",t.size),""}}});gt.restoreConversation();const t0={async listSessions(){const t=await fetch("api/chat/sessions/");if(!t.ok)throw new Error(`HTTP ${t.status}`);return(await t.json()).sessions??[]},async getMessages(t){const e=await fetch(`api/chat/sessions/${t}/messages/`);if(!e.ok)throw new Error(`HTTP ${e.status}`);return e.json()},async createSession(t){const e=await fetch("api/chat/sessions/",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({title:t||"New Chat"})});if(!e.ok)throw new Error(`HTTP ${e.status}`);return e.json()},async deleteSession(t){const e=await fetch(`api/chat/sessions/${t}/`,{method:"DELETE"});if(!e.ok)throw new Error(`HTTP ${e.status}`)},async addMessage(t,e,n){const s=await fetch(`api/chat/sessions/${t}/messages/`,{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({role:e,content:n})});if(!s.ok)throw new Error(`HTTP ${s.status}`)},async renameSession(t,e){const n=await fetch(`api/chat/sessions/${t}/`,{method:"PATCH",headers:{"Content-Type":"application/json"},body:JSON.stringify({title:e})});if(!n.ok)throw new Error(`HTTP ${n.status}`)}},Eu=new Bx,ku=document.getElementById("stx-shell-ai-sessions-list");ku&&(Eu.init(ku,t0,(t,e)=>{gt.loadSessionMessages(t.map(n=>({role:n.role,text:n.content,tools_used:[],media:[]})))},()=>{gt.clearChat()}),gt.setSessionsPanel(Eu));const yt=document.getElementById("stx-shell-ai-input");yt&&(yt.addEventListener("keydown",t=>{if(t.key==="Enter"&&!t.shiftKey){t.preventDefault(),gt.send();return}if((t.key==="ArrowUp"||t.ctrlKey&&t.key==="p")&&yt.selectionStart===0){t.preventDefault(),gt.navigateHistory(1);return}if((t.key==="ArrowDown"||t.ctrlKey&&t.key==="n")&&yt.selectionEnd===yt.value.length){t.preventDefault(),gt.navigateHistory(-1);return}}),yt.addEventListener("input",()=>{yt.style.height="auto",yt.style.height=Math.min(yt.scrollHeight,150)+"px"}));document.addEventListener("stx-shell:clear-chat",()=>{gt.clearChat()});window.addEventListener("stx-shell:camera",()=>{const t=document.querySelector('.stx-shell-ai-input-btn[title="Camera"]');t==null||t.click()});window.addEventListener("stx-shell:sketch",()=>{const t=document.querySelector('.stx-shell-ai-input-btn[title="Sketch"]');t==null||t.click()});window.addEventListener("stx-shell:mic-toggle",()=>{gt.toggleRecording()});let bu=!1;window.addEventListener("stx-shell:settings",()=>{if(bu)return;bu=!0;const t=document.getElementById("ai-agent-sources-content");t&&(t.innerHTML=`
     <div class="ai-config-category expanded" data-cat="App Skills">
       <div class="ai-config-category-header">
         <i class="fas fa-chevron-right ai-config-category-chevron"></i>
